% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acknowledge_packages.R
\name{acknowledge_packages}
\alias{acknowledge_packages}
\title{Acknowledge R packages.}
\usage{
acknowledge_packages(x, prefix = "We thank the ", suffix = " R packages.")
}
\arguments{
\item{x}{\code{character} names of package.}

\item{prefix}{\code{character} prefix text.
Defaults to \code{"We thank: "}.}

\item{suffix}{\code{character} suffix text.
Defaults to \code{" R packages."}.}
}
\value{
\code{shiny.tag} object.
}
\description{
This function creates a HTML element for acknowledging R packages.
}
