% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_acknowledgmentsSidebarPane.R
\name{acknowledgmentsSidebarPane}
\alias{acknowledgmentsSidebarPane}
\title{Acknowledgments sidebar pane}
\usage{
acknowledgmentsSidebarPane(id)
}
\arguments{
\item{id}{\code{character} identifier for the sidebar pane.}
}
\value{
A \code{shiny.tag} object with the sidebar pane.
}
\description{
Constructs a sidebar pane for displaying acknowledgments.
}
\details{
This is designed to be used as an argument to
\link[leaflet.extras2:sidebar_tabs]{leaflet.extras2::sidebar_tabs} when specifying
the user interface for a Shiny web application.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("allSidebars")
}
}

}
