% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leaflet_map_components.R
\name{add_leaflet_map_components}
\alias{add_leaflet_map_components}
\title{Add map components}
\usage{
add_leaflet_map_components(x, bbox = NULL)
}
\arguments{
\item{x}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} or \link[leaflet:leafletProxy]{leaflet::leafletProxy} object.}

\item{`list`}{of bounding box information.
Defaults to \code{NULL} such that no bounding box information is used.}
}
\value{
Updated version of argument to \code{x}.
}
\description{
Add components to a Leaflet map.
}
