% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_solutionResults_server.R
\name{addSolutionResults}
\alias{addSolutionResults}
\title{Add a solution results}
\usage{
addSolutionResults(session = shiny::getDefaultReactiveDomain(), inputId, value)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\link{Solution} object.}
}
\description{
Add a solution to the \link{solutionResults} widget.
}
