% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{all_list_elements_inherit}
\alias{all_list_elements_inherit}
\title{All elements inherit}
\usage{
all_list_elements_inherit(x, class)
}
\arguments{
\item{x}{\code{list} object.}

\item{class}{\code{character} class name.}
}
\value{
\code{logical} indicating if all elements in \code{x} inherit from \code{class}.
}
\description{
Check if all elements in a \code{list} object inherit from a class.
}
