% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_misc.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate coverage}
\usage{
calculate_coverage(x, data)
}
\arguments{
\item{x}{\code{numeric} solution values.}

\item{data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.}
}
\value{
\code{numeric} vector.
}
\description{
Calculate how well a solution covers data.
}
\examples{
# load dependency
library(Matrix)

# simulate solution values for 10 planning units
solution_values <- sample(c(0, 1), 10, replace = TRUE)

# simulate data for 5 features in those 10 planning units
feature_data <- as(matrix(runif(10 * 5), ncol = 10), "dgCMatrix")

# calculate coverage
calculate_coverage(solution_values, feature_data)
}
