% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_importSettings_server.R
\name{cleanImportSettings}
\alias{cleanImportSettings}
\title{Clean import settings widget}
\usage{
cleanImportSettings(session = shiny::getDefaultReactiveDomain(), inputId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}
}
\description{
Remove all field/layers displayed in the widget.
}
\seealso{
\code{\link[=importSettings]{importSettings()}}.
}
