% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_palette.R
\name{color_palette}
\alias{color_palette}
\title{Color palette}
\usage{
color_palette(x, n = NULL)
}
\arguments{
\item{x}{\code{character} name of the color palette
(see Details for available options).}

\item{n}{\code{integer} number of colors required.
A \code{NULL} argument can also be used to obtain a default number of colors.
If a \code{NULL} argument is supplied and the palette has a limited number of
colors, then the maximum number of colors is returned.
Otherwise, if a \code{NULL} argument is supplied and the palette does not have
a limited number of colors, then a 20 colors are returned.
The default argument is \code{NULL}.}
}
\description{
Generate colors using the name of a color palette.
}
\details{
This function can be used to generate colors using palettes provided by
the \pkg{RColorBrewer} and \pkg{viridisLite} packages. If colors are
from multiple palettes, they can be supplied as a single \code{character}
object delimited by semicolon (\code{";"}) characters.

Specifically, available palettes include:
\itemize{
\item \code{"viridis"} (see \code{\link[viridisLite:viridis]{viridisLite::viridis()}})
\item \code{"magma"} (see \code{\link[viridisLite:viridis]{viridisLite::magma()}})
\item \code{"plasma"} (see \code{\link[viridisLite:viridis]{viridisLite::plasma()}})
\item \code{"inferno"} (see \code{\link[viridisLite:viridis]{viridisLite::inferno()}})
\item \code{"cividis"} (see \code{\link[viridisLite:viridis]{viridisLite::cividis()}})
\item \code{"mako"} (see \code{\link[viridisLite:viridis]{viridisLite::mako()}})
\item \code{"rocket"} (see \code{\link[viridisLite:viridis]{viridisLite::rocket()}})
\item \code{"turbo"} (see \code{\link[viridisLite:viridis]{viridisLite::turbo()}})
\item \verb{"BrBG}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PiYG}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PRGn}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PuOr}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"RdBu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"RdGy}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"RdYlBu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"RdYlGn}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Spectral}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Accent}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Dark2}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Paired}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Pastel1}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Pastel2}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Set1}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Set2}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Set3}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Blues}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"BuGn}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"BuPu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"GnBu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Greens}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Greys}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Oranges}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"OrRd}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PuBu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PuBuGn}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"PuRd}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Purples}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"RdPu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"Reds}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"YlGn}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"YlGnBu}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"YlOrBr}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \verb{"YlOrRd}" (see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}})
\item \code{"random"} (the above palettes are randomly used to generate colors)
}
}
\examples{
# obtain 5 colors from the Greens palette
color_palette("Greens", 5)

# obtain all colors from the Greens palette
color_palette("Greens", NULL)

# obtain 15 colors using the Set1 and Pastel1 palettes
color_palette("Set1;Pastel1", 15)

# obtain all colors from the plasma palette
color_palette("plasma", NULL)

# obtain all colors from a random palette
color_palette("random", NULL)

# obtain 5 colors from random palette(s)
color_palette("random", 5)
}
