% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Dataset class}
\description{
Dataset class

Dataset class
}
\details{
Definition for the Dataset class.

Note that this method has no effect if the dataset does not have
file path on disk.
}
\seealso{
\code{\link[=new_dataset]{new_dataset()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}

\item{\code{spatial_data}}{\code{NULL}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[terra:rast]{terra::rast()}} object.}

\item{\code{attribute_data}}{\code{NULL}, or \code{\link[tibble:tibble]{tibble::tibble()}} object.}

\item{\code{boundary_data}}{\code{NULL}, or \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} or \code{NA} object.}

\item{\code{planning_unit_areas}}{\code{NULL}, or \code{numeric} vector of cached
planning unit areas, see \code{get_planning_unit_areas()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dataset-new}{\code{Dataset$new()}}
\item \href{#method-Dataset-print}{\code{Dataset$print()}}
\item \href{#method-Dataset-repr}{\code{Dataset$repr()}}
\item \href{#method-Dataset-import}{\code{Dataset$import()}}
\item \href{#method-Dataset-write}{\code{Dataset$write()}}
\item \href{#method-Dataset-clean}{\code{Dataset$clean()}}
\item \href{#method-Dataset-get_spatial_data}{\code{Dataset$get_spatial_data()}}
\item \href{#method-Dataset-get_attribute_data}{\code{Dataset$get_attribute_data()}}
\item \href{#method-Dataset-get_boundary_data}{\code{Dataset$get_boundary_data()}}
\item \href{#method-Dataset-get_crs}{\code{Dataset$get_crs()}}
\item \href{#method-Dataset-get_bbox}{\code{Dataset$get_bbox()}}
\item \href{#method-Dataset-get_planning_unit_indices}{\code{Dataset$get_planning_unit_indices()}}
\item \href{#method-Dataset-get_names}{\code{Dataset$get_names()}}
\item \href{#method-Dataset-get_planning_unit_areas}{\code{Dataset$get_planning_unit_areas()}}
\item \href{#method-Dataset-get_index}{\code{Dataset$get_index()}}
\item \href{#method-Dataset-has_index}{\code{Dataset$has_index()}}
\item \href{#method-Dataset-max_index}{\code{Dataset$max_index()}}
\item \href{#method-Dataset-add_index}{\code{Dataset$add_index()}}
\item \href{#method-Dataset-update_bm}{\code{Dataset$update_bm()}}
\item \href{#method-Dataset-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-new}{}}}
\subsection{Method \code{new()}}{
Create a Dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(
  id,
  spatial_path,
  attribute_path,
  boundary_path,
  spatial_data,
  attribute_data,
  boundary_data
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}

\item{\code{spatial_data}}{\code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[terra:rast]{terra::rast()}} object.}

\item{\code{attribute_data}}{\code{\link[tibble:tibble]{tibble::tibble()}} object.}

\item{\code{boundary_data}}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object or \link{NA}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Dataset object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-import"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-import}{}}}
\subsection{Method \code{import()}}{
Import the data into memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-write"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$write(spatial_path, attribute_path, boundary_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-clean"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-clean}{}}}
\subsection{Method \code{clean()}}{
Clean the dataset from memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_spatial_data"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_spatial_data}{}}}
\subsection{Method \code{get_spatial_data()}}{
Get the spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_spatial_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_attribute_data"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_attribute_data}{}}}
\subsection{Method \code{get_attribute_data()}}{
Get the attribute data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_attribute_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_boundary_data"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_boundary_data}{}}}
\subsection{Method \code{get_boundary_data()}}{
Get the spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_boundary_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_crs"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_crs}{}}}
\subsection{Method \code{get_crs()}}{
Get the coordinate reference system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_crs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_crs]{sf::st_crs()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_bbox}{}}}
\subsection{Method \code{get_bbox()}}{
Get the bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_bbox(native = TRUE, expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{native}}{\code{logical} indicating if the bounding box should
be in (\code{TRUE}) the native coordinate reference system or (\code{FALSE})
re-projected to longitude/latitude?}

\item{\code{expand}}{\code{FALSE} should the bounding box be expanded by 10\%?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} object with \code{"xmin"}, \code{"xmax"}, \code{"ymin"}, and \code{"ymax"}
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_planning_unit_indices"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_planning_unit_indices}{}}}
\subsection{Method \code{get_planning_unit_indices()}}{
Get planning unit indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_planning_unit_indices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} vector of indices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_names}{}}}
\subsection{Method \code{get_names()}}{
Get attribute names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector of field/layer names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_planning_unit_areas"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_planning_unit_areas}{}}}
\subsection{Method \code{get_planning_unit_areas()}}{
Get area values.
These are calculated once and then cached.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_planning_unit_areas()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector of values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-get_index}{}}}
\subsection{Method \code{get_index()}}{
Get a data from the dataset at an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-has_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-has_index}{}}}
\subsection{Method \code{has_index()}}{
Check if the dataset has an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$has_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical} indicating if data is present or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-max_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-max_index}{}}}
\subsection{Method \code{max_index()}}{
Maximum index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$max_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} largest index.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-add_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-add_index}{}}}
\subsection{Method \code{add_index()}}{
Add data at an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$add_index(index, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}

\item{\code{values}}{\code{numeric} vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-update_bm"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-update_bm}{}}}
\subsection{Method \code{update_bm()}}{
Updates old boundary matrix to new format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$update_bm()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{Matrix::sparseMatrix()} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
