% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs_functions.R
\name{disable_html_css_selector}
\alias{disable_html_css_selector}
\title{Disable HTML element(s) using a CSS selector}
\usage{
disable_html_css_selector(x)
}
\arguments{
\item{x}{\code{character} CSS selector.}
}
\value{
Invisible \code{TRUE}.

Invisible \code{TRUE}.
}
\description{
Disable HTML element(s) using a CSS selector.
}
