% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMapManagerLayer.R
\name{dropMapManagerLayer}
\alias{dropMapManagerLayer}
\title{Drop a layer from a map manager widget}
\usage{
dropMapManagerLayer(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\code{character} layer identifier.}
}
\description{
Drop a layer from a map manager widget on the client.
}
\seealso{
\code{\link[=mapManager]{mapManager()}}.
}
