% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_solutionResults_server.R
\name{dropSolutionResults}
\alias{dropSolutionResults}
\title{Drop a solution results}
\usage{
dropSolutionResults(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\code{character} identifier for the \link{Solution} to drop.}
}
\description{
Drop a solution to the \link{solutionResults} widget.
}
