% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinyjs.R
\name{enable_html_css_selector}
\alias{enable_html_css_selector}
\title{Enable a HTML element using a CSS selector}
\usage{
enable_html_css_selector(x)
}
\arguments{
\item{x}{\code{character} CSS selector.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Enable a HTML element using a CSS selector.
}
