% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinyjs.R
\name{enable_html_element}
\alias{enable_html_element}
\title{Enable a HTML element}
\usage{
enable_html_element(inputId)
}
\arguments{
\item{inputId}{HTML element id.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Enable a HTML element.
}
