% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_error_misc.R
\name{error_log}
\alias{error_log}
\title{Error log}
\usage{
error_log(x)
}
\arguments{
\item{x}{\code{try-error} object.}
}
\value{
\code{character} value.
}
\description{
Create an error log based on a \code{try-error} object.
}
\details{
This function checks \code{attr(x, "log")} for a list of \code{SimpleError} objects
that are used to construct the log. If the argument to \code{x} does not
contain this attribute, then the text \code{"No error log available."}
is returned.
}
