% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{error_message}
\alias{error_message}
\title{Error message}
\usage{
error_message(x)
}
\arguments{
\item{x}{\code{try-error} object.}
}
\value{
\code{character} value.
}
\description{
Create an error message based on a \code{try-error} object.
}
