% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExcludeResults.R
\name{ExcludeResults}
\alias{ExcludeResults}
\title{Exclude results class}
\description{
Exclude results class

Exclude results class
}
\details{
Definition for the \code{ExcludeResults} class.
}
\seealso{
\code{\link[=new_exclude_results]{new_exclude_results()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{exclude}}{\link{Exclude} object.}

\item{\code{status}}{\code{logical} value.}

\item{\code{held}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExcludeResults$new()}}
\item \href{#method-print}{\code{ExcludeResults$print()}}
\item \href{#method-repr}{\code{ExcludeResults$repr()}}
\item \href{#method-get_results_data}{\code{ExcludeResults$get_results_data()}}
\item \href{#method-get_widget_data}{\code{ExcludeResults$get_widget_data()}}
\item \href{#method-clone}{\code{ExcludeResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ExcludeResults} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$new(id, exclude, held)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{exclude}}{\link{Exclude} object.}

\item{\code{held}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ExcludeResults} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_results_data}{}}}
\subsection{Method \code{get_results_data()}}{
Get results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$get_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the object in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcludeResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
