% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_matrix.R
\name{extract_data_matrix}
\alias{extract_data_matrix}
\title{Extract data matrix}
\usage{
extract_data_matrix(x)
}
\arguments{
\item{x}{\code{list} of \link{Variable} objects.}
}
\value{
A \code{matrix} object.
}
\description{
Generate a \code{matrix} object containing data.
}
\details{
Note that all variables must have the same \link{Dataset} object.
}
