% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_projects.R
\name{find_projects}
\alias{find_projects}
\title{Find projects}
\usage{
find_projects(x, user_groups = "public")
}
\arguments{
\item{x}{\code{character} directory path.}

\item{user_groups}{\code{character} vector of project group names available that
can be imported.
Defaults to \code{"public"}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} object containing the configuration files
and names of projects found in the directory.
}
\description{
Find project files in a directory.
}
\details{
Projects that are missing user group information are automatically
assigned to the \code{"private"} user group.
}
\examples{
# find directory with built-in projects
d <- system.file("extdata", "projects", package = "wheretowork")

# list projects in directory
find_projects(d)
}
