% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_min_shortfall_solution.R
\name{generate_min_shortfall_solution}
\alias{generate_min_shortfall_solution}
\title{Generate minimum shortfall solution}
\usage{
generate_min_shortfall_solution(
  name,
  dataset,
  settings,
  area_budget_proportion,
  theme_data = settings$get_theme_data(),
  weight_data = settings$get_weight_data(),
  include_data = settings$get_include_data(),
  boundary_data = dataset$get_boundary_data(),
  gap = 0,
  boundary_budget_proportion = 0.1,
  legend_color = "#FF0000"
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{dataset}{\link{Dataset} object to store the solution.}

\item{settings}{\link{SolutionSettings} object with settings.}

\item{area_budget_proportion}{\code{numeric} budget for the solution.
This should be a proportion (with values ranging between 0 and 1)
indicating the maximum spatial extent of the solution.}

\item{theme_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the theme data.
Defaults to being automatically calculated from \code{settings}.}

\item{weight_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the weight data.
Defaults to being automatically calculated from \code{settings}.}

\item{include_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the include data.
Defaults to being automatically calculated from \code{settings}.}

\item{boundary_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the boundary data.
Defaults to being automatically extracted from \code{dataset}.}

\item{gap}{\code{numeric} relative optimality gap value. Defaults to 0.}

\item{boundary_budget_proportion}{\code{numeric} gap value used to control
the level of spatial clustering in the solution. Defaults to 0.1}

\item{legend_color}{\code{character} legend color.}
}
\value{
A \link{Solution} object containing the solution.
}
\description{
Create a new \link{Solution} object by generating a prioritization
using the minimum shortfall formulation of the reserve selection problem.
}
\examples{
#TODO.

}
