% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Read App Config}
\usage{
get_golem_config(
  value,
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE
)
}
\arguments{
\item{value}{Value to retrieve from the config file.}

\item{config}{GOLEM_CONFIG_ACTIVE value. If unset, R_CONFIG_ACTIVE.
If unset, "default".}

\item{use_parent}{Logical, scan the parent directory for config file.}
}
\description{
Read App Config
}
