% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontalPickerInput.R
\name{horizontalPickerInput}
\alias{horizontalPickerInput}
\title{Horizontal picker input}
\usage{
horizontalPickerInput(...)
}
\arguments{
\item{...}{arguments passed to \link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput}.}
}
\value{
A select list control that can be added to a UI definition.
}
\description{
Create a horizontal \link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput}.
}
