% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(x, mode)
}
\arguments{
\item{x}{\code{list} with project data.}

\item{mode}{\code{character} Mode for importing projects.
Available options include: \code{"beginner"}, \code{"advanced"}, or \code{"project"}.
Defaults to \code{"project"}.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Import project data into the app.
}
