% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModal.R
\name{importModal}
\alias{importModal}
\title{Import modal}
\usage{
importModal(id)
}
\arguments{
\item{id}{\code{character} identifier.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Constructs a modal for importing data.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("importModal")
}
}

}
