% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_importSettings_ui.R
\name{importSettings-shiny}
\alias{importSettings-shiny}
\alias{importSettingsOutput}
\alias{renderImportSettings}
\title{Shiny bindings for \code{importSettings}}
\usage{
importSettingsOutput(outputId, width = "100\%", height = "auto")

renderImportSettings(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a \code{\link[=importSettings]{importSettings()}}}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{importSettingsOutput()} to create a user interface element,
and \code{renderImportSettings()} to render the widget.
}
