% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSettings.R
\name{importSettings}
\alias{importSettings}
\title{Import settings}
\usage{
importSettings(
  buttonId,
  x = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{buttonId}{\code{character} containing the identifier for the button HTML
element to accompany this widget.}

\item{x}{\code{character} vector containing the field/layer names for the
dataset. A \code{NULL} value may also be specified to initialize the widget
without any field/layer names.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for specifying settings for importing data.
This widget is designed to be used within a Shiny web application.
}
\section{Server value}{

A \code{list} object containing a \code{list} element for each field/layer name
containing:
\describe{
\item{name}{\code{character} name of the field/layer.}
\item{import}{\code{logical} value indicating if the field/layer should be used.}
\item{type}{\code{character} value. Available options are
\code{"theme"}, \code{"weight"}, or \code{"include"}.}
}
}

\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("importSettings")
}
}

}
