% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Include.R
\name{Include}
\alias{Include}
\title{Include class}
\description{
Include class

Include class
}
\details{
Definition for the Include class.
}
\seealso{
\code{\link[=new_include]{new_include()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} name.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{overlap}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Include$new()}}
\item \href{#method-print}{\code{Include$print()}}
\item \href{#method-repr}{\code{Include$repr()}}
\item \href{#method-get_layer_name}{\code{Include$get_layer_name()}}
\item \href{#method-get_layer_id}{\code{Include$get_layer_id()}}
\item \href{#method-get_layer_index}{\code{Include$get_layer_index()}}
\item \href{#method-get_layer_pane}{\code{Include$get_layer_pane()}}
\item \href{#method-get_visible}{\code{Include$get_visible()}}
\item \href{#method-get_invisible}{\code{Include$get_invisible()}}
\item \href{#method-get_loaded}{\code{Include$get_loaded()}}
\item \href{#method-get_hidden}{\code{Include$get_hidden()}}
\item \href{#method-get_status}{\code{Include$get_status()}}
\item \href{#method-get_overlap}{\code{Include$get_overlap()}}
\item \href{#method-get_data}{\code{Include$get_data()}}
\item \href{#method-set_new_pane}{\code{Include$set_new_pane()}}
\item \href{#method-get_setting}{\code{Include$get_setting()}}
\item \href{#method-set_visible}{\code{Include$set_visible()}}
\item \href{#method-set_invisible}{\code{Include$set_invisible()}}
\item \href{#method-set_loaded}{\code{Include$set_loaded()}}
\item \href{#method-set_status}{\code{Include$set_status()}}
\item \href{#method-set_setting}{\code{Include$set_setting()}}
\item \href{#method-get_solution_settings_widget_data}{\code{Include$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Include$get_map_manager_widget_data()}}
\item \href{#method-export}{\code{Include$export()}}
\item \href{#method-render_on_map}{\code{Include$render_on_map()}}
\item \href{#method-update_on_map}{\code{Include$update_on_map()}}
\item \href{#method-clone}{\code{Include$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Include object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$new(
  id,
  name,
  variable,
  pane,
  invisible,
  loaded,
  mandatory,
  visible,
  hidden,
  status,
  overlap
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{overlap}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Include object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_name}{}}}
\subsection{Method \code{get_layer_name()}}{
Get layer names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_layer_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_id}{}}}
\subsection{Method \code{get_layer_id()}}{
Get layer id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_layer_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_index"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_index}{}}}
\subsection{Method \code{get_layer_index()}}{
Get layer index values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_layer_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_pane"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_pane}{}}}
\subsection{Method \code{get_layer_pane()}}{
Get layer pane class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_layer_pane()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-get_invisible}{}}}
\subsection{Method \code{get_invisible()}}{
Get invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} date/time value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-get_loaded}{}}}
\subsection{Method \code{get_loaded()}}{
Get loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-get_hidden}{}}}
\subsection{Method \code{get_hidden()}}{
Get hidden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_hidden()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-get_overlap}{}}}
\subsection{Method \code{get_overlap()}}{
Get overlap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_overlap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_new_pane"></a>}}
\if{latex}{\out{\hypertarget{method-set_new_pane}{}}}
\subsection{Method \code{set_new_pane()}}{
Set new pane.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_new_pane(id, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} unique identifier.}

\item{\code{index}}{\code{character} variable index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} or \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-set_invisible}{}}}
\subsection{Method \code{set_invisible()}}{
Set invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_invisible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{date/time} value or \code{NA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-set_loaded}{}}}
\subsection{Method \code{set_loaded()}}{
Set loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_loaded(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} or `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Include$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
