% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_validation.R
\name{is_valid_attribute_file}
\alias{is_valid_attribute_file}
\title{Is valid attribute file?}
\usage{
is_valid_attribute_file(x)
}
\arguments{
\item{x}{\code{character} file path.}
}
\value{
\code{logical} value indicating validity.
}
\description{
Verify if a file path corresponds to a valid attribute dataset.
}
