% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_file.R
\name{is_valid_boundary_file}
\alias{is_valid_boundary_file}
\title{Is valid boundary file?}
\usage{
is_valid_boundary_file(x)
}
\arguments{
\item{x}{\code{character} file path.}
}
\value{
\code{logical} value indicating validity.
}
\description{
Verify if a file path corresponds to a valid boundary dataset.
}
