% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer-class.R
\name{Layer}
\alias{Layer}
\title{Layer class}
\description{
Layer class

Layer class
}
\details{
Definition for the Layer class.
}
\seealso{
\code{\link[=new_layer]{new_layer()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{\code{character} value.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Layer$new()}}
\item \href{#method-print}{\code{Layer$print()}}
\item \href{#method-repr}{\code{Layer$repr()}}
\item \href{#method-clone}{\code{Layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Layer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$new(source, total, units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{\code{character} value.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Layer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
