% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{locationmisc}
\alias{locationmisc}
\title{locationmisc: Classes, Widgets, and Functions for the Location App}
\description{
This package provides classes, widgets, and functions for building
the Location App.
}
