% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManualLegend.R
\name{ManualLegend}
\alias{ManualLegend}
\title{Manual legend class}
\description{
Manual legend class

Manual legend class
}
\details{
Definition for the \code{ManualLegend} class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{\code{values} vector.}

\item{\code{colors}}{\code{character} vector.}

\item{\code{labels}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ManualLegend$new()}}
\item \href{#method-get_resample_method}{\code{ManualLegend$get_resample_method()}}
\item \href{#method-get_color_map}{\code{ManualLegend$get_color_map()}}
\item \href{#method-get_widget_data}{\code{ManualLegend$get_widget_data()}}
\item \href{#method-export}{\code{ManualLegend$export()}}
\item \href{#method-clone}{\code{ManualLegend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{ManualLegend} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$new(values, colors, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{\code{numeric} vector of values.}

\item{\code{colors}}{\code{character} vector of colors.}

\item{\code{labels}}{\code{character} vector of labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ManualLegend} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_resample_method"></a>}}
\if{latex}{\out{\hypertarget{method-get_resample_method}{}}}
\subsection{Method \code{get_resample_method()}}{
Get resample method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$get_resample_method()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_color_map"></a>}}
\if{latex}{\out{\hypertarget{method-get_color_map}{}}}
\subsection{Method \code{get_color_map()}}{
Get a function for mapping values to colors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$get_color_map()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{function} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for creating a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManualLegend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
