% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManager.R
\name{mapManager-shiny}
\alias{mapManager-shiny}
\alias{mapManagerOutput}
\alias{renderMapManager}
\title{Shiny bindings for mapManager}
\usage{
mapManagerOutput(outputId, width = "100\%", height = "auto")

renderMapManager(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a \code{\link[=mapManager]{mapManager()}}}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{mapManagerOutput()} to create a user interface element,
and \code{renderMapManager()} to render the widget.
}
