% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManager.R
\name{mapManager}
\alias{mapManager}
\title{Map manager}
\usage{
mapManager(x, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{x}{MapManager object.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for managing items on a Leaflet map.
This widget is designed to be used in conjunction with an existing
Leaflet Map within a Shiny web application.
}
\section{Server value}{

The widget sends the following values to the server
(where \code{elementId} corresponds to the \code{elementId}):

\describe{

\item{TODO}{\code{TODO}}

}
}

