% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManagerSidebarPane.R
\name{mapManagerSidebarPane}
\alias{mapManagerSidebarPane}
\title{Map manager sidebar pane}
\usage{
mapManagerSidebarPane(id, mapManagerId = paste0(id, "_settings"))
}
\arguments{
\item{id}{\code{character} identifier for the sidebar pane.}

\item{mapManagerId}{\code{character} identifier for the
\code{\link[=mapManager]{mapManager()}} widget to create within the sidebar pane.
This widget is used to control the settings for new solutions.
Defaults to \code{paste0(id, "_settings")}.}
}
\value{
A \code{shiny.tag} object with the sidebar pane.
}
\description{
Constructs a sidebar pane for managing the layers on a
\code{\link[leaflet:map-shiny]{leaflet::leafletOutput()}} map.
}
\details{
This is designed to be used as an argument to
\link[leaflet.extras2:sidebar_tabs]{leaflet.extras2::sidebar_tabs} when specifying
the user interface for a Shiny web application.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("mapManagerSidebarPane")
}
}

}
