% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_min_set_solution.R
\name{min_set_solution}
\alias{min_set_solution}
\title{Generate minimum set solution}
\usage{
min_set_solution(
  name,
  dataset,
  settings,
  theme_data = settings$get_theme_data(),
  weight_data = settings$get_weight_data(),
  include_data = settings$get_include_data(),
  boundary_data = dataset$get_boundary_data(),
  gap = 0,
  boundary_gap = 0.1,
  legend_color = "#FF0000"
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{dataset}{\link{Dataset} object to store the solution.}

\item{settings}{\link{SolutionSettings} object with settings.}

\item{theme_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the theme data.
Defaults to being automatically calculated from \code{settings}.}

\item{weight_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the weight data.
Defaults to being automatically calculated from \code{settings}.}

\item{include_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the include data.
Defaults to being automatically calculated from \code{settings}.}

\item{boundary_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} containing the boundary data.
Defaults to being automatically extracted from \code{dataset}.}

\item{gap}{\code{numeric} relative optimality gap value. Defaults to 0.}

\item{boundary_gap}{\code{numeric} gap value used to control
the level of spatial clustering in the solution. Defaults to 0.1}

\item{legend_color}{\code{character} legend color.}
}
\value{
A \link{Solution} object containing the solution.
}
\description{
Create a new \link{Solution} object by generating a prioritization
using the minimum set formulation of the reserve selection problem.
}
\examples{
# TODO.
}
