% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTheme-class.R
\name{MultiTheme}
\alias{MultiTheme}
\title{MultiTheme class}
\description{
MultiTheme class

MultiTheme class
}
\details{
Definition for the MultiTheme class.
}
\seealso{
\code{\link[=new_multi_theme]{new_multi_theme()}}.
}
\section{Super class}{
\code{\link[locationmisc:Theme]{locationmisc::Theme}} -> \code{MultiTheme}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature_order}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiTheme$new()}}
\item \href{#method-get_feature_order}{\code{MultiTheme$get_feature_order()}}
\item \href{#method-set_feature_order}{\code{MultiTheme$set_feature_order()}}
\item \href{#method-get_solution_settings_widget_data}{\code{MultiTheme$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{MultiTheme$get_map_manager_widget_data()}}
\item \href{#method-clone}{\code{MultiTheme$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="export">}\href{../../locationmisc/html/Theme.html#method-export}{\code{locationmisc::Theme$export()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_current">}\href{../../locationmisc/html/Theme.html#method-get_feature_current}{\code{locationmisc::Theme$get_feature_current()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_goal">}\href{../../locationmisc/html/Theme.html#method-get_feature_goal}{\code{locationmisc::Theme$get_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_id">}\href{../../locationmisc/html/Theme.html#method-get_feature_id}{\code{locationmisc::Theme$get_feature_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_name">}\href{../../locationmisc/html/Theme.html#method-get_feature_name}{\code{locationmisc::Theme$get_feature_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_status">}\href{../../locationmisc/html/Theme.html#method-get_feature_status}{\code{locationmisc::Theme$get_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_total">}\href{../../locationmisc/html/Theme.html#method-get_feature_total}{\code{locationmisc::Theme$get_feature_total()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_visible">}\href{../../locationmisc/html/Theme.html#method-get_feature_visible}{\code{locationmisc::Theme$get_feature_visible()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_layer_index">}\href{../../locationmisc/html/Theme.html#method-get_layer_index}{\code{locationmisc::Theme$get_layer_index()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_layer_name">}\href{../../locationmisc/html/Theme.html#method-get_layer_name}{\code{locationmisc::Theme$get_layer_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_setting">}\href{../../locationmisc/html/Theme.html#method-get_setting}{\code{locationmisc::Theme$get_setting()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="print">}\href{../../locationmisc/html/Theme.html#method-print}{\code{locationmisc::Theme$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="render_on_map">}\href{../../locationmisc/html/Theme.html#method-render_on_map}{\code{locationmisc::Theme$render_on_map()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="repr">}\href{../../locationmisc/html/Theme.html#method-repr}{\code{locationmisc::Theme$repr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_goal">}\href{../../locationmisc/html/Theme.html#method-set_feature_goal}{\code{locationmisc::Theme$set_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_status">}\href{../../locationmisc/html/Theme.html#method-set_feature_status}{\code{locationmisc::Theme$set_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_visible">}\href{../../locationmisc/html/Theme.html#method-set_feature_visible}{\code{locationmisc::Theme$set_feature_visible()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_setting">}\href{../../locationmisc/html/Theme.html#method-set_setting}{\code{locationmisc::Theme$set_setting()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="update_on_map">}\href{../../locationmisc/html/Theme.html#method-update_on_map}{\code{locationmisc::Theme$update_on_map()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a MultiTheme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$new(id, name, feature, mandatory, feature_order, icon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{feature_order}}{\code{numeric} vector.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MultiTheme object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_order}{}}}
\subsection{Method \code{get_feature_order()}}{
Set relative order for displaying features on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_feature_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_order}{}}}
\subsection{Method \code{set_feature_order()}}{
Set relative order for displaying features on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$set_feature_order(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector of new orders.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
