% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MultiThemeResults.R
\name{MultiThemeResults}
\alias{MultiThemeResults}
\title{MultiThemeResults class}
\description{
MultiThemeResults class

MultiThemeResults class
}
\details{
Definition for the MultiThemeResults class.
}
\seealso{
\code{\link[=new_theme_results]{new_theme_results()}}.
}
\section{Super class}{
\code{\link[wheretowork:ThemeResults]{wheretowork::ThemeResults}} -> \code{MultiThemeResults}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_results_data}{\code{MultiThemeResults$get_results_data()}}
\item \href{#method-get_widget_data}{\code{MultiThemeResults$get_widget_data()}}
\item \href{#method-clone}{\code{MultiThemeResults$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="wheretowork" data-topic="ThemeResults" data-id="initialize">}\href{../../wheretowork/html/ThemeResults.html#method-initialize}{\code{wheretowork::ThemeResults$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wheretowork" data-topic="ThemeResults" data-id="print">}\href{../../wheretowork/html/ThemeResults.html#method-print}{\code{wheretowork::ThemeResults$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wheretowork" data-topic="ThemeResults" data-id="repr">}\href{../../wheretowork/html/ThemeResults.html#method-repr}{\code{wheretowork::ThemeResults$repr()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_results_data}{}}}
\subsection{Method \code{get_results_data()}}{
Get results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiThemeResults$get_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the object in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiThemeResults$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiThemeResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
