% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_CategoricalLegend.R
\name{new_categorical_legend}
\alias{new_categorical_legend}
\title{New categorical legend}
\usage{
new_categorical_legend(values, colors)
}
\arguments{
\item{values}{\code{numeric} Values to show in the legend.}

\item{colors}{\code{character} Colors to show in the legend.
These colors should be in hex format (e.g. \code{"#AABBCC"}).}
}
\value{
A \link{CategoricalLegend} object.
}
\description{
Create a new \link{CategoricalLegend} object.
}
\examples{
# create new object
l <- new_categorical_legend(c(0, 1, 2), c("#000000", "#444444", "#AAAAAA"))

# print object
print(l)

}
