% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{new_dataset}
\alias{new_dataset}
\title{New dataset}
\usage{
new_dataset(source, total, units)
}
\arguments{
\item{source}{\code{character} file path for the underlying data.}

\item{total}{\code{numeric} total amount of all values in the underlying data.}

\item{units}{\code{character} units for the values in the underlying data.}
}
\value{
A \link{Dataset} object.
}
\description{
Create a new \link{Dataset} object.
}
\examples{
# create new object
l <- new_dataset(source = tempfile(), total = 12, units = "ha")

# print object
print(l)

}
