% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Exclude.R
\name{new_exclude}
\alias{new_exclude}
\title{New exclude}
\usage{
new_exclude(
  name,
  variable,
  mandatory = FALSE,
  visible = TRUE,
  hidden = FALSE,
  status = FALSE,
  overlap = NA_character_,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{variable}{\link{Variable} object.}

\item{mandatory}{\code{logical} value indicating if object is mandatory
for generating solutions.}

\item{visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}.}

\item{hidden}{\code{logical} The hidden value.
This is used to determine if the feature is can ever be displayed (or not)
or not the map. Unlike \code{visible}, if this parameter is \code{FALSE} then a
feature can never be viewed on the map.
Defaults to \code{FALSE}.}

\item{status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}.}

\item{overlap}{\code{character} vector that remains \code{NA_character_} until set in
the \link{SolutionSettings} object. \code{overlap} defines the include and exclude
overlap.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Exclude} object.
}
\description{
Create a new \link{Exclude} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new exclude
e <- new_exclude(name = "Protected areas", variable = v)

# print object
print(e)
}
