% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExcludeResults.R
\name{new_exclude_results}
\alias{new_exclude_results}
\title{New exclude results}
\usage{
new_exclude_results(exclude, held, id = uuid::UUIDgenerate())
}
\arguments{
\item{exclude}{\link{Weight} object.}

\item{held}{\code{numeric} proportion of the exclude covered by the solution.
(e.g. 0.1 = 10\%).}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{ExcludeResults} object.
}
\description{
Create a new \link{ExcludeResults} object to store results for a solution.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new exclude
excl <- new_exclude(name = "Highways", variable = v)

# create a new exclude results object to store results
exclr <- new_exclude_results(excl, 80)

# print object
print(exclr)
}
