% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-class.R
\name{new_feature}
\alias{new_feature}
\title{New feature}
\usage{
new_feature(
  name,
  dataset,
  initial_visible = TRUE,
  initial_status = TRUE,
  initial_goal = 0.1,
  min_goal = 0,
  max_goal = 1,
  step_goal = 0.01,
  limit_goal = 0.1,
  current = 0,
  current_label = "Current",
  icon = "map-marked-alt",
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name of the feature.}

\item{dataset}{\link{Dataset} object.}

\item{initial_visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}}

\item{initial_status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{initial_goal}{\code{numeric} The initial goal for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0.3 (i.e. 30\%).}

\item{min_goal}{\code{numeric} The minimum goal (inclusive)
shown for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0 (i.e. 0\%).}

\item{max_goal}{\code{numeric} The maximum goal (inclusive) shown
for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 1 (i.e. 100\%).}

\item{step_goal}{\code{numeric} The minimum increment for setting goals.
Note that goal are specified as proportions, such that a
value of 0.01 corresponds to 1\%.
Defaults to 0.01 (i.e. 1\%).}

\item{limit_goal}{\code{numeric} The minimum goal
(inclusive) that can be selected for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0 (i.e. 0\%).}

\item{current}{\code{numeric} current proportion of values held in existing
conservation areas (e.g. 0.1 = 10\%).}

\item{current_label}{\code{character} The display label for the current
level of for the feature.
Defaults to \code{"Current"}.}

\item{icon}{\code{character} or \code{shiny.tag} Icon to display for the feature
This icon should indicate the type of data that underpin the feature.
If the argument to \code{icon} is a \code{character}, it is used with
\code{\link[shiny:icon]{shiny::icon()}} to generate an \code{shiny.tag} icon.
Defaults to \code{"map-marked-alt"}.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Feature} object.
}
\description{
Create a new \link{Feature} object.
}
\examples{
# create new dataset
l <- new_dataset(
 source = tempfile(), total = 12, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#AAAAAA")))

# create feature using the dataset
f <- new_feature(name = "Intact Alvar", dataset = l)

# print object
print(f)

}
