% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Include-class.R
\name{new_include}
\alias{new_include}
\title{New include}
\usage{
new_include(
  name,
  variable,
  mandatory = FALSE,
  initial_visible = TRUE,
  initial_status = TRUE,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{variable}{\link{Variable} object.}

\item{initial_visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}}

\item{initial_status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Include} object.
}
\description{
Create a new \link{Include} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new include
w <- new_include(name = "Protected areas", variable = v)

# print object
print(w)

}
