% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_IncludeResults.R
\name{new_include_results}
\alias{new_include_results}
\title{New include results}
\usage{
new_include_results(include, held, id = uuid::UUIDgenerate())
}
\arguments{
\item{include}{\link{Weight} object.}

\item{held}{\code{numeric} proportion of the include covered by the solution.
(e.g. 0.1 = 10\%).}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{IncludeResults} object.
}
\description{
Create a new \link{IncludeResults} object to store results for a solution.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new include
i <- new_include(name = "NDVI", variable = v)

# create a new include results object to store results
ir <- new_include_results(i, 80)

# print object
print(ir)
}
