% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer-class.R
\name{new_layer}
\alias{new_layer}
\title{New layer}
\usage{
new_layer(source, current, total, units)
}
\arguments{
\item{source}{\code{character} file path for the underlying data.}

\item{current}{\code{numeric} current proportion of values held in existing
conservation areas (e.g. 0.1 = 10\%).}

\item{total}{\code{numeric} total amount of all values in the underlying data.}

\item{units}{\code{character} units for the values in the underlying data.}
}
\value{
A \link{Layer} object.
}
\description{
Create a new \link{Layer} object.
}
\examples{
# create new object
l <- new_layer(source = tempfile(), current = 0.1, total = 12, units = "ha")

# print object
print(l)

}
