% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualLegend-class.R
\name{new_manual_legend}
\alias{new_manual_legend}
\title{New manual legend}
\usage{
new_manual_legend(colors, labels)
}
\arguments{
\item{colors}{\code{character} Colors to show in the legend.
These colors should be in hex format (e.g. \code{"#AABBCC"}).
Arguments should contain two different colors.}

\item{labels}{\code{character} Labels to show in the legend.}
}
\value{
A \link{ManualLegend} object.
}
\description{
Create a new \link{ManualLegend} object.
}
\examples{
# create new object
l <- new_manual_legend(c("#000000", "#AAAAAA"), c("a", "b"))

# print object
print(l)

}
