% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapManager-class.R
\name{new_map_manager}
\alias{new_map_manager}
\title{New map manager}
\usage{
new_map_manager(
  layers,
  visible = rep(TRUE, length(layers)),
  order = seq_along(layers)
)
}
\arguments{
\item{layers}{\code{list} of \link{Theme} and/or \link{Weight} objects.}
}
\value{
A \link{MapManager} object.
}
\description{
Create a new \link{MapManager} object.
}
\examples{
# create datasets
l1 <- new_dataset(
 source = tempfile(), total = 12, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#1b9e77")))
l2 <- new_dataset(
 source = tempfile(), total = 14, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#d95f02")))
l3 <- new_dataset(
 source = tempfile(), total = 78, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#7570b3")))
l4 <- new_dataset(
 source = tempfile(), total = 90, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#e31a1c")))

# create a weight using a dataset
w <- new_weight(name = "Human Footprint Index", dataset = l1)

# create features using datasets
f1 <- new_feature(name = "Possum Occurrence", dataset = l2)
f2 <- new_feature(name = "Forests", dataset = l3)
f3 <- new_feature(name = "Shrublands", dataset = l4)

# create themes using the features
t1 <- new_single_theme(name = "Species", f1)
t2 <- new_multi_theme(name = "Ecoregions", list(f1, f2))

# create a map manager for the themes and weight
mm <- new_map_manager(layers = list(t1, t2, w))

# print object
print(mm)

}
