% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTheme-class.R
\name{new_multi_theme}
\alias{new_multi_theme}
\title{New multi-theme}
\usage{
new_multi_theme(
  name,
  feature,
  mandatory = FALSE,
  round = TRUE,
  icon = "map-marked-alt",
  current_label = "Current",
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{feature}{\code{list} of \link{Feature} objects.}

\item{mandatory}{\code{logical} Is the theme mandatory for generating solutions?
Defaults to \code{FALSE}.}

\item{round}{\code{logical} should all numbers be rounded to the nearest integer?
Defaults to \code{TRUE}.}

\item{icon}{\code{shiny.tag} Icon to display for the feature
This icon should indicate the type of data that underpin the feature.
Alternatively, the argument can be a \code{character} to automatically
generate a \code{shiny.tag} icon (using \code{\link[shiny:icon]{shiny::icon()}}).
Defaults to \code{"map-marked-alt"}.}

\item{current_label}{\code{character} The display label for the current
level of representation for the features under the group view.
Defaults to \code{Current}.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{MultiTheme} object.
}
\description{
Create a new \link{MultiTheme} object.
}
\examples{
# create datasets
l1 <- new_dataset(
 source = tempfile(), total = 12, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#1b9e77")))
l2 <- new_dataset(
 source = tempfile(), total = 14, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#d95f02")))
l3 <- new_dataset(
 source = tempfile(), total = 78, units = "ha",
 legend = new_continuous_legend(1, 100, c("#000000", "#7570b3")))

# create features the datasets
f1 <- new_feature(name = "Pangolin", dataset = l1)
f2 <- new_feature(name = "Panda", dataset = l2)
f3 <- new_feature(name = "Palms", dataset = l3)

# create a multi-theme using the features
mt <- new_multi_theme(
  name = "Endangered species", feature = list(f1, f2, f3))

# print object
print(mt)

}
