% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter-class.R
\name{new_parameter}
\alias{new_parameter}
\title{New parameter}
\usage{
new_parameter(
  name,
  initial_status = TRUE,
  initial_value = 0,
  min_value = 0,
  max_value = 100,
  step_value = 1,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{initial_status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{initial_value}{\code{numeric} initial value.
Defaults to 0.}

\item{min_value}{\code{numeric} minimum value.
Defaults to 0.}

\item{max_value}{\code{numeric} maximum value.}

\item{step_value}{\code{numeric} step value.
Defaults to 1.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Parameter} object.
}
\description{
Create a new \link{Parameter} object.
}
\examples{
# create a new parameter
p <- new_parameter(name = "Spatial clumping")

# print object
print(p)

}
