% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Provenance.R
\name{new_provenance}
\alias{new_provenance}
\title{New Provenance}
\usage{
new_provenance(name, icon, color, description)
}
\arguments{
\item{name}{\code{character} name of provenance source.}

\item{icon}{\code{character} name of icon to depict the soruce.}

\item{color}{\code{character} value containing the color for the icon.}

\item{description}{\code{character} value  containing a description of the
provenance.}
}
\value{
A \link{Provenance} object.
}
\description{
Create a new \link{Provenance} object.
}
\examples{
# create a provenance
x <- new_provenance(
  name = "Global", icon = "globe", color = "#123456",
  description = "Data is global."
)

# print object
print(x)
}
