% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Provenance.R
\name{new_provenance_from_source}
\alias{new_provenance_from_source}
\title{New provenance from source}
\usage{
new_provenance_from_source(source)
}
\arguments{
\item{source}{\code{character} provenance type.
Available options include \code{"national"}, \code{"regional"}, or \code{"missing"}.}
}
\value{
A \link{Provenance} object.
}
\description{
Create a new \link{Provenance} object from a recognized source.
}
\examples{
# create a provenance
x <- new_provenance_from_source("missing")

# print object
print(x)

}
