% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{new_solution}
\alias{new_solution}
\title{New solution}
\usage{
new_solution(
  name,
  variable,
  initial_visible,
  statistics,
  theme_results,
  weight_results,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{variable}{\link{Variable} object with the solution.}

\item{initial_visible}{\code{logical} should the solution be visible on a map?}

\item{statistics}{\code{list} of \link{Statistic} objects.}

\item{theme_results}{\code{list} of \link{ThemeResults} objects.}

\item{weight_results}{\code{list} of \link{WeightResults} objects.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\description{
Create a new \link{Solution} object.
}
\examples{
#TODO

}
