% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution-class.R
\name{new_solution_from_prioritization}
\alias{new_solution_from_prioritization}
\title{New solution from analysis}
\usage{
new_solution_from_prioritization(
  name,
  dataset,
  solution_settings,
  rij,
  wij,
  boundary_data,
  gap = 0,
  boundary_penalty_gap = 0.1
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{dataset}{\link{Dataset} object to store the solution.}

\item{solution_settings}{\link{SolutionSettings} object.}

\item{rij}{\code{matrix} object.
This matrix specifies the amount of each feature associated with
each planning unit.}

\item{wij}{\code{matrix} object.
This matrix specifies the value of each weight associated with each
planning unit.}

\item{boundary_data}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.
This matrix specifies the boundary length data for the planning units.}

\item{gap}{\code{numeric} relative optimality gap value. Defaults to 0.}

\item{boundary_penalty_gap}{\code{numeric} gap value used to control
the level of spatial fragmentation in the solution. Defaults to 0.1}
}
\value{
A \code{SolutionResults} object containing the solution.
}
\description{
Create a new \link{Solution} object by generating a prioritization and
calculating performance statistics.
}
\examples{
#TODO.

}
