% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{new_solution_from_result}
\alias{new_solution_from_result}
\title{New solution from result}
\usage{
new_solution_from_result(
  name,
  visible,
  dataset,
  settings,
  result,
  legend,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{visible}{\code{logical} should the solution be visible on a map?}

\item{dataset}{\link{Dataset} object.}

\item{settings}{\link{SolutionSettings} object.}

\item{result}{\link{Result} object.}

\item{legend}{\link{ManualLegend} object.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Solution} object.
}
\description{
Create a new \link{Solution} object using a \link{Result} object
}
\examples{
#TODO
}
