% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionLegend-class.R
\name{new_solution_legend}
\alias{new_solution_legend}
\title{New solution legend}
\usage{
new_solution_legend(colors)
}
\arguments{
\item{colors}{\code{character} Colors to show in the legend.
These colors should be in hex format (e.g. \code{"#AABBCC"}).
Arguments should contain two different colors.}
}
\value{
A \link{SolutionLegend} object.
}
\description{
Create a new \link{SolutionLegend} object.
}
\examples{
# create new object
l <- new_solution_legend(c("#000000", "#AAAAAA"))

# print object
print(l)

}
