% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionSettings-class.R
\name{new_solution_settings}
\alias{new_solution_settings}
\title{New solution settings}
\usage{
new_solution_settings(themes, weights)
}
\arguments{
\item{themes}{\code{list} of \link{Theme} objects. D}

\item{weights}{\code{list} of \link{Weight} objects.}
}
\value{
A \link{SolutionSettings} object.
}
\description{
Create a new \link{SolutionSettings} object.
}
\examples{
# create datasets
l1 <- new_dataset(source = tempfile(), total = 12, units = "ha")
l2 <- new_dataset(source = tempfile(), total = 14, units = "ha")
l3 <- new_dataset(source = tempfile(), total = 78, units = "ha")
l4 <- new_dataset(source = tempfile(), total = 90, units = "ha")

# create a weight using a dataset
w <- new_weight(name = "Human Footprint Index", dataset = l1)

# create features using datasets
f1 <- new_feature(name = "Possum Occurrence", dataset = l2)
f2 <- new_feature(name = "Forests", dataset = l3)
f3 <- new_feature(name = "Shurblands", dataset = l4)

# create themes using the features
t1 <- new_single_theme(name = "Species", f1)
t2 <- new_multi_theme(name = "Ecoregions", list(f1, f2))

# create solution settings using the themes and weight
ss <- new_solution_settings(themes = list(t1, t2), weights = list(w))

# print object
print(ss)

}
