% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeResults-class.R
\name{new_theme_results}
\alias{new_theme_results}
\title{New theme results}
\usage{
new_theme_results(theme, feature_results, id = uuid::UUIDgenerate())
}
\arguments{
\item{theme}{\link{Theme} object.}

\item{feature_results}{\link{FeatureResults} object or a \code{list} of such objects.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{ThemeResults} object.
}
\description{
Create a new \link{ThemeResults} object to store results for a solution.
}
\examples{
#TODO

}
