% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Variable.R
\name{new_variable}
\alias{new_variable}
\title{New variable}
\usage{
new_variable(
  dataset,
  index,
  units,
  total,
  legend,
  provenance = new_provenance_from_source("missing")
)
}
\arguments{
\item{dataset}{\code{Dataset} file path for the dataset.}

\item{index}{\code{character} or \code{integer} indicating the field/layer with
the data.}

\item{units}{\code{character} units for the values in the underlying data.}

\item{total}{\code{numeric} total amount of all values in the underlying data.}

\item{legend}{\code{Legend} object.}

\item{provenance}{\link{Provenance} object.
Defaults to \code{new_provenance_from_source("missing")}.}
}
\value{
A \link{Variable} object.
}
\description{
Create a new \link{Variable} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable(d,
  index = 1, total = 12, units = "ha",
  legend = new_continuous_legend(1, 100, c("#000000", "#AAAAAA"))
)

# print object
print(v)
}
