% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Variable.R
\name{new_variable_from_auto}
\alias{new_variable_from_auto}
\title{New variable from automatic calculations}
\usage{
new_variable_from_auto(
  dataset,
  index,
  units = "",
  type = "auto",
  colors = "random",
  provenance = "missing",
  labels = "missing",
  hidden = FALSE
)
}
\arguments{
\item{dataset}{\code{Dataset} file path for the dataset.}

\item{index}{\code{character} or \code{integer} indicating the field/layer with
the data.}

\item{units}{\code{character} units for the values in the underlying data.}

\item{type}{\code{character} indicating if the data contain
continuous (\code{"continuous"}) or categorical (\code{"categorical"})
numerical values.
Defaults to \code{"auto"} such that data are automatically identified.}

\item{colors}{\code{character} object containing the colors for visualization
(see Details for more information).
Defaults to \code{"random"} such that colors are randomly generated.}

\item{provenance}{\code{character} value indicating the type of provenance.
The argument must be a valid type (see \code{\link[=new_provenance_from_source]{new_provenance_from_source()}}).
Defaults to \code{"missing"}.}

\item{labels}{\code{character} object containing manual legend labels.
Defaults to \code{"missing"}.}

\item{hidden}{\code{logical} indicating if variable will be hidden from map.
Defaults to \code{"FALSE"}. If hidden, a \link{NullLegend} is used}
}
\value{
A \link{Variable} object.
}
\description{
Create a new \link{Variable} object by automatically calculating
all metadata from the underlying data.
This function is useful when pre-calculated metadata are not available.
Note this function will take longer to create variables than other
functions because it requires performing geospatial operations.
}
\details{
The argument to \code{colors} can be a vector of different colors
(in hexadecimal format, e.g. \verb{"#112233"), or a single }character\verb{containing the name of a color palette that is used to generate a vector of different colors (see [color_palette()] for more information). The color palette name}"random"` is also available, such that
colors are generated using a randomly selected palette.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# print object
print(v)
}
