% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable-class.R
\name{new_variable_from_metadata}
\alias{new_variable_from_metadata}
\title{New variable from metadata}
\usage{
new_variable_from_metadata(dataset, metadata)
}
\arguments{
\item{dataset}{\code{Dataset} file path for the dataset.}

\item{metadata}{\code{list} object (see Details for more information).}
}
\description{
Create a new \link{Variable} object using metadata.
This function is useful when pre-calculated are available, so that
previously calculated metadata can be used.
}
\details{
The argument to \code{metadata} should contain the following elements:

\describe{
\item{"index"}{\code{character} or \code{integer} indicating the field/layer
with the data within the dataset that has the data.}
\item{"units"}{\code{character} units for the values in the underlying data.}
\item{"type"}{\code{character} indicating if the data contain
continuous (\code{"continuous"}) or categorical (\code{"categorical"})
numerical values.}
\item{"colors"}{\code{character} vector containing colors for visualization.}
\item{"total"}{\code{numeric} sum of all values in dataset.}
\item{"min_value"}{\code{numeric} minimum value in dataset.
Required only for continuous data.}
\item{"max_value"}{\code{numeric} maximum value in dataset.
Required only for continuous data.}
\item{"values"}{\code{numeric} vector of unique value in dataset.
Required only for categorical data.}
}
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_metadata(
  d, list(index = 1, units = "ha", type = "continuous",
          colors = c("#000000", "#AAAAAA"), total = 12,
          min_value = 1, max_value = 3))

# print object
print(v)

}
