% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight-class.R
\name{new_weight}
\alias{new_weight}
\title{New weight}
\usage{
new_weight(
  name,
  layer,
  min_factor = 0,
  max_factor = 100,
  initial_factor = 0,
  step_factor = 1,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} name.}

\item{layer}{\link{Layer} object.}

\item{min_factor}{\code{numeric} minimum factor value.
Defaults to 0.}

\item{max_factor}{\code{numeric} maximum factor value.}

\item{initial_factor}{\code{numeric} initial factor value.
Defaults to 0.}

\item{step_factor}{\code{numeric} step factor value.
Defaults to 1.}

\item{id}{\code{character} identifier.
Defaults to a unique identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Weight} object.
}
\description{
Create a new \link{Weight} object.
}
\examples{
# create a new layer
l <- new_layer(source = tempfile(), total = 12, units = "ha")

# create a new weight
w <- new_weight(name = "NDVI", layer = l)

# print object
print(w)

}
