% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightResults-class.R
\name{new_weight_results}
\alias{new_weight_results}
\title{New weight results}
\usage{
new_weight_results(weight, held, id = uuid::UUIDgenerate())
}
\arguments{
\item{weight}{\link{Weight} object.}

\item{held}{\code{numeric} proportion of the weight covered by the solution.
(e.g. 0.1 = 10\%).}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{WeightResults} object.
}
\description{
Create a new \link{WeightResults} object to store results for a solution.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new weight
w <- new_weight(name = "NDVI", variable = v)

# create a new weight results object to store results
wr <- new_weight_results(w, 80)

# print object
print(w)

}
