% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSolutionManager.R
\name{newSolutionManager}
\alias{newSolutionManager}
\title{New solution manager}
\usage{
newSolutionManager(
  themes,
  weights,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{themes}{\code{list} of \code{Theme} objects to include in the manager.}

\item{weights}{\code{list} of \code{Weight} objects to include in the manager.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for generating new solutions and managing their
associated settings.
}
\section{Server value}{

The widget sends the following values to the server
(where \code{eId} corresponds to the \code{elementId}):

\describe{

\item{eIdrun}{\code{character} name to use for a new solution.
This value is updated when the user clicks the button to generate
a new solution.}

\item{eId_setting}{\code{list} containing updated settings.
This value is updated when the user updates the settings
(e.g. status, factor, goal) for a \link{Weight} or \link{Theme}.
Specifically, the \code{list} contains the following elements:

\describe{
\item{id}{\code{character} identifier for the theme or weight.}
\item{parameter}{\code{character} name of the updated parameter.
Available options include: \code{"status"}, \code{"factor"}, or "goal".}
\item{value}{new \code{numeric} or \code{logical} values.}
\item{type}{\code{character} indicating if the updated parameter corresponds
to a \code{theme} or \code{weight}.}
}
}
}
}

\examples{
# TODO.

}
