% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_NullLegend.R
\name{NullLegend}
\alias{NullLegend}
\title{NULL legend class}
\description{
NULL legend class

NULL legend class
}
\details{
Definition for the \code{NullLegend} class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{\code{numeric} vector.}

\item{\code{colors}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NullLegend$new()}}
\item \href{#method-get_widget_data}{\code{NullLegend$get_widget_data()}}
\item \href{#method-clone}{\code{NullLegend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{NullLegend} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLegend$new(values, colors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{\code{NULL} value.}

\item{\code{colors}}{\code{NULL} vector of colors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{NullLegend} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for creating a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLegend$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{CategoricalLegend} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLegend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
