% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fileInput.R
\name{prepare_fileInput}
\alias{prepare_fileInput}
\title{Prepare file input}
\usage{
prepare_fileInput(x)
}
\arguments{
\item{x}{\code{data.frame} object.}
}
\value{
\code{character} file paths.
}
\description{
Prepare files uploaded using \link[shiny:fileInput]{shiny::fileInput}.
}
\details{
This function renames files to their original file names after
they have been uploaded.
}
