% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Provenance.R
\name{Provenance}
\alias{Provenance}
\title{Provenance class}
\description{
Provenance class

Provenance class
}
\details{
Definition for the Provenance class.
}
\seealso{
\code{\link[=new_provenance]{new_provenance()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{\code{character} value.}

\item{\code{icon}}{\code{character} value}

\item{\code{color}}{\code{character} value.}

\item{\code{description}}{\code{character} value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Provenance-new}{\code{Provenance$new()}}
\item \href{#method-Provenance-print}{\code{Provenance$print()}}
\item \href{#method-Provenance-repr}{\code{Provenance$repr()}}
\item \href{#method-Provenance-get_widget_data}{\code{Provenance$get_widget_data()}}
\item \href{#method-Provenance-export}{\code{Provenance$export()}}
\item \href{#method-Provenance-clone}{\code{Provenance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-new}{}}}
\subsection{Method \code{new()}}{
Create a Provenance object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$new(name, icon, color, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} value.}

\item{\code{icon}}{\code{numeric} value}

\item{\code{color}}{\code{character} value.}

\item{\code{description}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Provenance object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-print"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get widget data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-export"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-export}{}}}
\subsection{Method \code{export()}}{
Get widget data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
