% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{read_dataset}
\alias{read_dataset}
\title{Read dataset}
\usage{
read_dataset(source, subset, type = "auto", colors = "random")
}
\arguments{
\item{source}{\code{character} file path for the dataset.}

\item{subset}{\code{character} or \code{integer} indicating the field/layer with
data within the dataset.}

\item{type}{\code{character} indicating if the data contain
continuous (\code{"continuous"}) or categorical (\code{"categorical"})
numerical values.
Defaults to \code{"auto"} such that data are automatically identified.}

\item{colors}{\code{character} object containing the colors for visualization
(see Details for more information).
Defaults to \code{"random"} such that colors are randomly generated.}
}
\value{
A \link{Dataset} object.
}
\description{
Read a \link{Dataset} object from disk
This function is useful when pre-calculated metadata are not available,
because it will automatically calculate the required metadata to import
the dataset.
}
\details{
The argument to \code{colors} can be a vector of different colors
(in hexadecimal format, e.g. \verb{"#112233"), or a single }character\verb{containing the name of a color palette that is used to generate a vector of different colors (see [color_palette()] for more information). The color palette name}"random"` is also available, such that
colors are generated using a randomly selected palette.
}
\examples{
# load example raster
data(sim_pu_raster, package = "prioritizr")
r <- sim_pu_raster

# create temporary file name for saving raster to disk
f <- tempfile(fileext = ".tif")

# save raster to disk
f <- writeRaster(r, f, NAflag = -9999)

# create new object
d <- read_dataset(f)

# print object
print(d)

}
