% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{read_dataset_with_metadata}
\alias{read_dataset_with_metadata}
\title{Read dataset with metadata}
\usage{
read_dataset_with_metadata(x)
}
\arguments{
\item{x}{\code{list} object (see Details for more information).}
}
\description{
Read a \link{Dataset} object from disk using metadata.
This function is useful when pre-calculated are available, so that
previously calculated metadata can be used to import the dataset.
}
\details{
The argument to \code{x} should contain the following elements:

\describe{
\item{"source"}{\code{character} file path for the data.}
\item{"subset"}{\code{character} or \code{integer} indicating the field/layer
with the data within the dataset.}
\item{"units"}{\code{character} units for the values in the underlying data.}
\item{"type"}{\code{character} indicating if the data contain
continuous (\code{"continuous"}) or categorical (\code{"categorical"})
numerical values.}
\item{"colors"}{\code{character} vector containing colors for visualization.}
\item{"statistics"}{\code{list} object containing statistics:
\describe{
\item{"total"}{\code{numeric} sum of all values in dataset.
Required for both continuous and categorical data.}
\item{"min_value"}{\code{numeric} minimum value in dataset.
Required for continuous data, and optional for categorical data.}
\item{"max_value"}{\code{numeric} maximum value in dataset.
Required for continuous data, and optional for categorical data.}
\item{"values"}{\code{numeric} vector of unique value in dataset.
Required for categorical data, and optional for continuous data.}
}
}
}
}
