% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_project.R
\name{read_project}
\alias{read_project}
\title{Read project}
\usage{
read_project(
  path,
  spatial_path = NULL,
  attribute_path = NULL,
  boundary_path = NULL,
  mode = "project",
  force_hidden = FALSE
)
}
\arguments{
\item{path}{\code{character} file path for the configuration file.}

\item{spatial_path}{\code{character} file path for spatial data.
Defaults to \code{NULL} such that the file path is determined using
the argument to \code{path}.}

\item{attribute_path}{\code{character} file path for attribute data.
Defaults to \code{NULL} such that the file path is determined using
the argument to \code{path}.}

\item{boundary_path}{\code{character} file path for boundary data.
Defaults to \code{NULL} such that the file path is determined using
the argument to \code{path}.}

\item{mode}{\code{character} mode for importing the objects.
Defaults to \code{"project"} such that the mode is determined based on
the contents of \code{path}. If the \code{mode} is \code{"advanced"}, then
goal limits and mandatory include settings are disabled.}

\item{force_hidden}{\code{logical} value for hiding TWI layers at import.
Defaults to \code{"False"} such that hidden TWI layers are determined based on
the contents of \code{path}.}
}
\value{
A \code{list} containing the following elements:
\describe{
\item{name}{A \code{character} value indicating the name.}
\item{dataset}{A \link{Dataset} object.}
\item{themes}{A \code{list} of \link{Theme} objects.}
\item{weights}{A \code{list} of \link{Weight} objects.}
\item{includes}{A \code{list} of \link{Include} objects.}
\item{mode}{A \code{character} value indicating the mode.}
\item{force_hidden}{A \code{logical} value indicating all TWI will be hidden.}
}
}
\description{
Read a project from disk.
}
\details{
Note that the status field for themes and weights will automatically
be set to \code{FALSE} if the goal/weight values are zero.
This is to ensure that the application starts with a sensible
combination of settings.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_data.yaml",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f4 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# read project
x <- read_project(f1, f2, f3, f4)

# print project
print(x)
}
