% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_io.R
\name{readNamedRaster}
\alias{readNamedRaster}
\title{Read a \code{Raster} object from disk}
\usage{
readNamedRaster(x, ...)
}
\arguments{
\item{x}{filename (character), Raster* object, missing (to create an empty RasterStack), SpatialGrid*, SpatialPixels*, or list (of filenames and/or Raster* objects). If \code{x} is a list, additional arguments \code{...} are ignored}

\item{...}{additional filenames or Raster* objects}
}
\value{
result from \code{\link[raster]{stack}} with names.
}
\description{
Read a \code{Raster} object from disk
}
\details{
This file reads a \code{\link{Raster}} object to disk along with
a text file containing the layer names of the raster.
}
