% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinyjs.R
\name{removeMapPane}
\alias{removeMapPane}
\title{Remove map pane using CSS selector}
\usage{
removeMapPane(id)
}
\arguments{
\item{id}{\code{character} CSS selector.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Remove map pane using CSS selector
}
