% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_misc.R
\name{repair_spatial_data}
\alias{repair_spatial_data}
\title{Repair spatial data}
\usage{
repair_spatial_data(x)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} object.}
}
\value{
\code{\link[sf:sf]{sf::st_sf()}} object.
}
\description{
Repair geometry in a spatial dataset.
}
\examples{
path <- system.file("shape/nc.shp", package = "sf")
g <- sf::read_sf(path)
repair_spatial_data(g)
}
