% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Result.R
\name{Result}
\alias{Result}
\title{Result class}
\description{
Result class

Result class
}
\details{
Definition for the Result class.
}
\seealso{
\code{\link[=new_result]{new_result()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{values}}{\code{numeric} vector.}

\item{\code{area}}{\code{numeric} value.}

\item{\code{perimeter}}{\code{numeric} value or \code{NA} logical.}

\item{\code{theme_coverage}}{\code{numeric} value.}

\item{\code{weight_coverage}}{\code{numeric} value.}

\item{\code{include_coverage}}{\code{numeric} value.}

\item{\code{theme_settings}}{\code{data.frame} object.}

\item{\code{weight_settings}}{\code{data.frame} object.}

\item{\code{include_settings}}{\code{data.frame} object.}

\item{\code{parameters}}{\code{list} of \link{Parameter} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Result$new()}}
\item \href{#method-print}{\code{Result$print()}}
\item \href{#method-repr}{\code{Result$repr()}}
\item \href{#method-clone}{\code{Result$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Result object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Result$new(
  id,
  values,
  area,
  perimeter,
  theme_coverage,
  weight_coverage,
  include_coverage,
  theme_settings,
  weight_settings,
  include_settings,
  parameters
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{values}}{\code{numeric} vector.}

\item{\code{area}}{\code{numeric} value.}

\item{\code{perimeter}}{\code{numeric} value.}

\item{\code{theme_coverage}}{\code{numeric} vector.}

\item{\code{weight_coverage}}{\code{numeric} vector.}

\item{\code{include_coverage}}{\code{numeric} vector.}

\item{\code{theme_settings}}{\code{logical} value.}

\item{\code{weight_settings}}{\code{logical} value.}

\item{\code{include_settings}}{\code{logical} value.}

\item{\code{parameters}}{\code{list} of \link{Parameter} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Result object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Result$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Result$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Result$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
