% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_misc.R
\name{rgb_raster_brick}
\alias{rgb_raster_brick}
\title{RGB raster}
\usage{
rgb_raster_brick(x, fun)
}
\arguments{
\item{x}{\code{RasterLayer} object.}

\item{fun}{\code{function} containing a color legend function
(e.g. \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}.}
}
\value{
\code{RasterBrick} object.
}
\description{
Create a RGB (red-blue-green) raster brick to render a raster layer based
on a color legend function.
}
\examples{
# create raster
f <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
r <- raster(f)
r[] <- runif(f, ncell(r))

# create legend function
l <- leaflet::colorNumeric(palette = "viridis", domain = c(0, 1))

# plot raster (using default color ramp)
plot(r)

# create RGB raster
r2 <- rgb_raster_brick(r, l)

# plot RGB raster
plotRGB(r2)

}
