% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples}
\usage{
runExample(name)
}
\arguments{
\item{name}{\code{character} name of widget.
Available options include \code{"solutionSettings"} and \code{"mapManager"}.}
}
\description{
Run a 'Shiny' application containing showcasing widgets implemented
in the \pkg{locationmisc} package.
}
\examples{
\dontrun{
if (interactive()) {
  runExample("solutionSettings")
}
}
}
