% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_misc.R
\name{shapefile_field_names}
\alias{shapefile_field_names}
\title{Shapefile field names}
\usage{
shapefile_field_names(x, inherits = NULL)
}
\arguments{
\item{x}{\code{character} path.}

\item{inherits}{\code{character} class to return only column names that
contain a certain type of data.
Defaults to \code{NULL} such that all column names are returned.}
}
\value{
\code{character} vector of field names.
}
\description{
Quickly extract the field names in a shapefile.
}
\examples{
shapefile_field_names(system.file("shape/nc.shp", package = "sf"))
}
