% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSolutionResults.R
\name{showSolutionResults}
\alias{showSolutionResults}
\title{Show solution results}
\usage{
showSolutionResults(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\code{character} identifier for the \link{Solution} to show.}
}
\description{
Show the results for a solution in the \link{solutionResults} widget.
}
