% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_legends.R
\name{simulate_categorical_legend}
\alias{simulate_categorical_legend}
\title{Simulate categorical legend}
\usage{
simulate_categorical_legend()
}
\value{
A \link{CategoricalLegend} object.
}
\description{
This function simulates a categorical legend object.
It is primarily used for testing the package.
}
\examples{
print(simulate_continuous_legend())
}
\seealso{
\code{\link[=new_categorical_legend]{new_categorical_legend()}}.
}
