% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_legends.R
\name{simulate_continuous_legend}
\alias{simulate_continuous_legend}
\title{Simulate continuous legend}
\usage{
simulate_continuous_legend()
}
\value{
A \link{ContinuousLgend} object.
}
\description{
This function simulates a continuous legend (\link{ContinuousLegend}) object.
It is primarily used for testing the package.
}
\examples{
print(simulate_continuous_legend())

}
\seealso{
\code{\link[=new_continuous_legend]{new_continuous_legend()}}.
}
