% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  n_single_themes = 5,
  n_multi_themes = 15,
  n_weights = 5,
  lambda = 5
)
}
\arguments{
\item{n_single_themes}{\code{integer} number of themes with a single feature
to simulate. Defaults to 5.}

\item{n_multi_themes}{\code{integer} number of themes with a multiple features
to simulate. Defaults to 15.}

\item{n_weights}{\code{integer} number of weights to simulate.
Defaults to 5.}

\item{lambda}{\code{numeric} lambda parameter for simulating the number
of features for each \link{MultiTheme} object.
This parameter is used as a argument to \code{rpois(n, lambda = lambda)}.
Defaults to 5.}
}
\value{
A \code{list} containing the simulated objects. It has \code{themes} and
\code{weights} elements that contain \link{Theme} and \link{Weight} objects (respectively).
}
\description{
This function simulates \link{Weight}, \link{SingleTheme}, and \link{MultiTheme} objects.
}
\examples{
# simulate data
x <- simulate_data(
  n_single_themes = 3, n_multi_themes = 2, n_weights = 1)

# print results
print(x)

}
\seealso{
\link{new_weight}, \link{new_single_theme}, \link{new_multi_theme}.
}
