% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_legends.R
\name{simulate_exclude_legend}
\alias{simulate_exclude_legend}
\title{Simulate exclude legend}
\usage{
simulate_exclude_legend()
}
\value{
A \link{ManualLegend} object.
}
\description{
This function simulates a legend for a \link{Exclude} object.
It is primarily used for testing the package.
}
\examples{
print(simulate_exclude_legend())
}
