% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_excludes.R
\name{simulate_excludes}
\alias{simulate_excludes}
\title{Simulate excludes}
\usage{
simulate_excludes(dataset, n)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{n}{\code{integer} number of objects to simulate.}
}
\value{
A \code{list} of simulated \link{Exclude} objects.
}
\description{
This function simulates \link{Exclude} objects.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
if (requireNamespace("RandomFields")) {
 x <- simulate_excludes(dataset = d, n = 2)
 # print results
 print(x)
}

}
\seealso{
\link{new_exclude}.
}
