% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_legends.R
\name{simulate_include_legend}
\alias{simulate_include_legend}
\title{Simulate include legend}
\usage{
simulate_include_legend()
}
\value{
A \link{ManualLegend} object.
}
\description{
This function simulates a legend for a \link{Include} object.
It is primarily used for testing the package.
}
\examples{
print(simulate_include_legend())
}
