% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spatial_dataset.R
\name{simulate_simple_raster_data}
\alias{simulate_simple_raster_data}
\title{Simulate simple raster data}
\usage{
simulate_simple_raster_data()
}
\value{
A \code{\link[raster:raster]{raster::raster()}} object.
}
\description{
Simulate a simple spatial raster dataset.
}
