% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spatial_dataset.R
\name{simulate_simple_vector_data}
\alias{simulate_simple_vector_data}
\title{Simulate simple vector data}
\usage{
simulate_simple_vector_data()
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}} object.
}
\description{
Simulate a simple spatial vector dataset.
}
