% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_solution.R
\name{simulate_solution}
\alias{simulate_solution}
\title{Simulate a new solution}
\usage{
simulate_solution(dataset, themes, weights)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{themes}{\code{list} of \link{Theme} objects.}

\item{weights}{\code{list} of \link{Weight} objects.}
}
\value{
A \link{Solution} object.
}
\description{
This function simulates a \link{Solution} object.
}
