% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_legends.R
\name{simulate_solution_legend}
\alias{simulate_solution_legend}
\title{Simulate solution legend}
\usage{
simulate_solution_legend()
}
\value{
A \link{SolutionLegend} object.
}
\description{
This function simulates a solution legend (\link{SolutionLegend}) object.
It is primarily used for testing the package.
}
\examples{
print(simulate_solution_legend())

}
\seealso{
\code{\link[=new_solution_legend]{new_solution_legend()}}.
}
