% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_solution_settings.R
\name{simulate_solution_settings}
\alias{simulate_solution_settings}
\title{Simulate a solution settings object}
\usage{
simulate_solution_settings(
  n_single_themes = 5,
  n_multi_themes = 15,
  n_weights = 5,
  lambda = 5
)
}
\arguments{
\item{n_single_themes}{\code{integer} number of themes with a single feature
to simulate. Defaults to 5.}

\item{n_multi_themes}{\code{integer} number of themes with a multiple features
to simulate. Defaults to 15.}

\item{n_weights}{\code{integer} number of weights to simulate.
Defaults to 5.}

\item{lambda}{\code{numeric} lambda parameter for simulating the number
of features for each theme.
This parameter is used as a argument to \code{rpois(n, lambda = lambda)}.
Defaults to 5.}
}
\value{
A \link{SolutionSettings} object.
}
\description{
This function simulates a solution settings object. It is primarily used
for testing the package.
}
\examples{
# simulate solution settings
x <- simulate_solution_settings(
  n_single_themes = 3, n_multi_themes = 2, n_weights = 1)

# print results
print(x)

}
\seealso{
\code{\link[=new_solution_settings]{new_solution_settings()}}.
}
