% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_themes.R
\name{simulate_themes}
\alias{simulate_themes}
\title{Simulate themes}
\usage{
simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{n_single_themes}{\code{integer} number of themes with a single feature
to simulate.}

\item{n_multi_themes}{\code{integer} number of themes with a multiple features
to simulate.}

\item{lambda}{\code{numeric} lambda parameter for simulating the number
of features for themes that contain multiple features.
This parameter is used as a argument to \code{rpois(n, lambda = lambda)}.
Defaults to 5.}
}
\value{
A \code{list} of simulated \link{Theme} objects.
}
\description{
This function simulates \link{Theme} objects.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
if (requireNamespace("RandomFields")) {
 x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
 # print results
 print(x)
}

# print results
print(x)
}
\seealso{
\link{new_theme}.
}
