% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_weights.R
\name{simulate_weights}
\alias{simulate_weights}
\title{Simulate weights}
\usage{
simulate_weights(dataset, n)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{n}{\code{integer} number of weights to simulate.}
}
\value{
A \code{list} of simulated \link{Weight} objects.
}
\description{
This function simulates \link{Weight} objects.
}
\examples{
# import data
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")
d <- new_dataset(raster::raster(f))

# simulate data
x <- simulate_weights(dataset = d, n = 2)

# print results
print(x)

}
\seealso{
\link{new_weight}.
}
