% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionLegend-class.R
\name{SolutionLegend}
\alias{SolutionLegend}
\title{SolutionLegend class}
\description{
SolutionLegend class

SolutionLegend class
}
\details{
Definition for the SolutionLegend class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colors}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolutionLegend$new()}}
\item \href{#method-get_resample_method}{\code{SolutionLegend$get_resample_method()}}
\item \href{#method-get_color_map}{\code{SolutionLegend$get_color_map()}}
\item \href{#method-get_widget_data}{\code{SolutionLegend$get_widget_data()}}
\item \href{#method-clone}{\code{SolutionLegend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a CategoricalLegend object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionLegend$new(colors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colors}}{\code{character} vector of colors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SolutionLegend object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_resample_method"></a>}}
\if{latex}{\out{\hypertarget{method-get_resample_method}{}}}
\subsection{Method \code{get_resample_method()}}{
Get resample method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionLegend$get_resample_method()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_color_map"></a>}}
\if{latex}{\out{\hypertarget{method-get_color_map}{}}}
\subsection{Method \code{get_color_map()}}{
Get a function for mapping values to colors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionLegend$get_color_map()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{function} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for creating a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionLegend$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new CategoricalLegend object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionLegend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
