% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionResults.R
\name{solutionResults-shiny}
\alias{solutionResults-shiny}
\alias{solutionResultsOutput}
\alias{renderSolutionResults}
\title{Shiny bindings for solutionResults}
\usage{
solutionResultsOutput(outputId, width = "100\%", height = "auto")

renderSolutionResults(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a \code{\link[=solutionResults]{solutionResults()}}}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{solutionResultsOutput()} to create a user interface element,
and \code{renderSolutionResults()} to render the widget.
}
