% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionResults.R
\name{solutionResults}
\alias{solutionResults}
\title{Solution results}
\usage{
solutionResults(x = list(), width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{x}{\code{list} containing \link{Solution} objects.
Defaults to an empty list object.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for displaying solution results.
This widget is designed to be used in conjunction with an existing
Leaflet Map within a Shiny web application.
}
\section{Server value}{

The widget does not send any server values.
}

\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("solutionResults")
}
}

}
