% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_solutionResultsModal.R
\name{solutionResultsModal}
\alias{solutionResultsModal}
\title{Solution results modal}
\usage{
solutionResultsModal(id, trigger)
}
\arguments{
\item{id}{\code{character} identifier for the modal.}

\item{trigger}{\code{character} identifier for the modal trigger.
See \link[shinyBS:bsModal]{shinyBS::bsModal} for further details.}
}
\value{
A \code{shiny.tag} object with the modal
}
\description{
Constructs a modal for displaying solution results.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("solutionResultsModal")
}
}

}
