% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_data_type.R
\name{spatial_data_type}
\alias{spatial_data_type}
\alias{spatial_data_type.sf}
\alias{spatial_data_type.SpatRaster}
\alias{spatial_data_type.data.frame}
\title{Spatial data type}
\usage{
spatial_data_type(x, index = 1, ...)

\method{spatial_data_type}{sf}(x, index = 1, ...)

\method{spatial_data_type}{SpatRaster}(x, index = 1, max_sample = 10000, ...)

\method{spatial_data_type}{data.frame}(x, index = 1, ...)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} or combined \code{\link[terra:rast]{terra::rast()}} or \link{data.frame} dataset object.}

\item{index}{\code{integer} or \code{character} value indicating the
field or layer for which to calculate statistics.
Defaults to 1, such that the first field/layer is used to calculate
statistics.}

\item{...}{not used.}

\item{max_sample}{\code{integer} maximum number of cells in a raster stack
to sample when checking if the data are continuous or categorical.
Defaults to 10000.}
}
\value{
A \code{character} indicating if the data are \code{"continuous"} or
\code{"categorical"}.
}
\description{
Identify if a spatial dataset has continuous or categorical data.
}
