% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{staticModal}
\alias{staticModal}
\title{Static modal}
\usage{
staticModal(id, title, trigger, ...)
}
\arguments{
\item{id}{\code{character} HTML element identifier.}

\item{...}{content.}
}
\description{
Create a static modal.
}
