% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Statistic.R
\name{Statistic}
\alias{Statistic}
\title{Statistic class}
\description{
Statistic class

Statistic class
}
\details{
Definition for the Statistic class.
}
\seealso{
\code{\link[=new_statistic]{new_statistic()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{\code{character} value.}

\item{\code{value}}{\code{numeric} value}

\item{\code{units}}{\code{character} value.}

\item{\code{proportion}}{\code{numeric} value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Statistic$new()}}
\item \href{#method-print}{\code{Statistic$print()}}
\item \href{#method-repr}{\code{Statistic$repr()}}
\item \href{#method-get_widget_data}{\code{Statistic$get_widget_data()}}
\item \href{#method-get_results_data}{\code{Statistic$get_results_data()}}
\item \href{#method-clone}{\code{Statistic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Statistic object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$new(name, value, units, proportion)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} value.}

\item{\code{value}}{\code{numeric} value}

\item{\code{units}}{\code{character} value.}

\item{\code{proportion}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Statistic object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get widget data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_results_data}{}}}
\subsection{Method \code{get_results_data()}}{
Get results data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$get_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
