% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme-class.R
\name{Theme}
\alias{Theme}
\title{Theme class}
\description{
Theme class

Theme class
}
\details{
Definition for the Theme class. This is a parent class that contains
shared fields and methods for the \link{SingleTheme} and \link{MultiTheme} classes.

This method intentionally throws an error.
}
\seealso{
\link{new_single_theme}, \link{new_multi_theme}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{status}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{round}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Theme$new()}}
\item \href{#method-get_feature_goals}{\code{Theme$get_feature_goals()}}
\item \href{#method-get_feature_statuses}{\code{Theme$get_feature_statuses()}}
\item \href{#method-set_feature_goals}{\code{Theme$set_feature_goals()}}
\item \href{#method-set_feature_statuses}{\code{Theme$set_feature_statuses()}}
\item \href{#method-clone}{\code{Theme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Theme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_goals"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_goals}{}}}
\subsection{Method \code{get_feature_goals()}}{
Get feature goals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_goals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector of goal values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_statuses"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_statuses}{}}}
\subsection{Method \code{get_feature_statuses()}}{
Get feature statuses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_statuses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector of status values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goals"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goals}{}}}
\subsection{Method \code{set_feature_goals()}}{
Set feature goals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_goals(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector of new values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_statuses"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_statuses}{}}}
\subsection{Method \code{set_feature_statuses()}}{
Set feature goals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_statuses(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector of new values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
