% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ThemeResults.R
\name{ThemeResults}
\alias{ThemeResults}
\title{Theme results class}
\description{
Theme results class

Theme results class
}
\details{
Definition for the \code{ThemeResults} class.
}
\seealso{
\code{\link[=new_theme_results]{new_theme_results()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{theme}}{\link{Theme} object.}

\item{\code{feature_results}}{\code{list} of \link{FeatureResults} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThemeResults-new}{\code{ThemeResults$new()}}
\item \href{#method-ThemeResults-print}{\code{ThemeResults$print()}}
\item \href{#method-ThemeResults-repr}{\code{ThemeResults$repr()}}
\item \href{#method-ThemeResults-get_results_data}{\code{ThemeResults$get_results_data()}}
\item \href{#method-ThemeResults-get_widget_data}{\code{ThemeResults$get_widget_data()}}
\item \href{#method-ThemeResults-clone}{\code{ThemeResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeResults} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$new(id, theme, feature_results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{theme}}{`\link{Theme} object.}

\item{\code{feature_results}}{\code{list} of \link{FeatureResults} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeResults} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-repr"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-get_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-get_results_data}{}}}
\subsection{Method \code{get_results_data()}}{
Get results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$get_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the object in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
