% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateImportSettings.R
\name{updateImportSettings}
\alias{updateImportSettings}
\title{Update import settings widget}
\usage{
updateImportSettings(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\code{character} Vector containing the new field/layers.}
}
\description{
Change the field/layers displayed in the widget.
}
\seealso{
\code{\link[=importSettings]{importSettings()}}.
}
