% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMapManagerOrder.R
\name{updateMapManagerOrder}
\alias{updateMapManagerOrder}
\title{Update map manager widget ordering}
\usage{
updateMapManagerOrder(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}

\item{inputId}{\code{character} The identifier of the input object.}

\item{value}{\code{numeric} Vector containing the new ordering.}
}
\description{
Change the order of the layers for a map manager widget on the client.
}
\seealso{
\code{\link[=mapManager]{mapManager()}}.
}
