% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useCustomScrollbars.R
\name{useCustomScrollbars}
\alias{useCustomScrollbars}
\title{Use custom scrollbars}
\usage{
useCustomScrollbars()
}
\description{
This function is used to add custom scrollbars to a Shiny application.
}
