% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wheretowork}
\alias{wheretowork}
\title{wheretowork: Systematic conservation planning application.}
\description{
The application is a decision support tool to help prioritize conservation
efforts for the Nature Conservancy of Canada. It provides an interactive
interface for conducting systematic conservation planning exercises, and
uses mathematical optimization algorithms to generate solutions.
}
\examples{
\donttest{
if (interactive()) {
run_app()
}
}
}
