% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_excel_workbook.R
\name{write_excel_workbook}
\alias{write_excel_workbook}
\title{Write Excel workbook}
\usage{
write_excel_workbook(x, path)
}
\arguments{
\item{x}{\code{data.frame} object (or \code{list} of \code{data.frame} objects) to save to
disk.}

\item{path}{\code{character} file path to save data.}
}
\value{
Invisible \code{TRUE} indicating success.
}
\description{
Save a table to disk as an Excel Spreadsheet.
}
\details{
This function is similar to the \link[openxlsx:write.xlsx]{openxlsx::write.xlsx} function.
The main difference is that it formats cells columns containing \code{numeric}
values (and \code{numeric}-like values) using the \code{"NUMBER"} format, and the
remaining cells as \code{"TEXT"} format.
}
\examples{
# create file path to save data
f <- tempfile(fileext = ".xlsx")

# load example dataset to save to disk
data(iris)
d <- iris

# save dataset to disk
write_excel_workbook(d, f)
}
