% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_io.R
\name{write_spatial_data}
\alias{write_spatial_data}
\title{Write spatial data}
\usage{
write_spatial_data(x, dir, name)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.}

\item{dir}{\code{character} directory to save data.}

\item{name}{\code{character} name of the spatial data file.}
}
\value{
Invisible \code{TRUE} indicating success.
}
\description{
Write spatial data to disk.
}
\examples{
# read and write raster data
f1 <- system.file("external/rlogo.grd", package = "raster")
d1 <- read_spatial_data(f1)
write_spatial_data(terra::rast(d1), tempdir(), "rlogo")

# read and write vector data
f2 <- system.file("shape/nc.shp", package = "sf")
d2 <- read_spatial_data(f2)
write_spatial_data(d2, tempdir(), "nc")

}
