% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_io.R
\name{writeNamedRaster}
\alias{writeNamedRaster}
\title{Write a \code{Raster} object to disk}
\usage{
writeNamedRaster(x, filename, ...)
}
\arguments{
\item{x}{Raster* object}

\item{filename}{Output filename}

\item{...}{arguments passed to \code{\link[raster]{writeRaster}}.}
}
\value{
result from  \code{\link[raster]{writeRaster}}.
}
\description{
Write a \code{Raster} object to disk
}
\details{
This file saves a \code{\link{Raster}} object to disk along with
a text file containing the layer names of the raster.
}
