% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalLegend-class.R
\name{CategoricalLegend}
\alias{CategoricalLegend}
\title{CategoricalLegend class}
\description{
CategoricalLegend class

CategoricalLegend class
}
\details{
Definition for the CategoricalLegend class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{\code{numeric} vector.}

\item{\code{colors}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CategoricalLegend$new()}}
\item \href{#method-get_widget_data}{\code{CategoricalLegend$get_widget_data()}}
\item \href{#method-export}{\code{CategoricalLegend$export()}}
\item \href{#method-clone}{\code{CategoricalLegend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a CategoricalLegend object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalLegend$new(values, colors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{\code{numeric} value.}

\item{\code{colors}}{\code{character} vector of colors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CategoricalLegend object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for creating a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalLegend$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new CategoricalLegend object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalLegend$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalLegend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
