% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinyjs.R
\name{change_file_icon_js}
\alias{change_file_icon_js}
\title{Display red X icon and update tool tip; fileinput_parameter}
\usage{
change_file_icon_js(x)
}
\arguments{
\item{x}{\code{character} CSS selector.}
}
\value{
Invisible \code{TRUE}.

Invisible \code{TRUE}.
}
\description{
Display red X icon and update tool tip; fileinput_parameter
}
