% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ContinuousLegend.R
\name{ContinuousLegend}
\alias{ContinuousLegend}
\title{Continuous legend class}
\description{
Continuous legend class

Continuous legend class
}
\details{
Definition for the \code{ContinuousLegend} class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{min_value}}{\code{numeric} value.}

\item{\code{max_value}}{\code{numeric} value.}

\item{\code{colors}}{\code{character} vector.}

\item{\code{n}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContinuousLegend-new}{\code{ContinuousLegend$new()}}
\item \href{#method-ContinuousLegend-get_widget_data}{\code{ContinuousLegend$get_widget_data()}}
\item \href{#method-ContinuousLegend-get_resample_method}{\code{ContinuousLegend$get_resample_method()}}
\item \href{#method-ContinuousLegend-get_color_map}{\code{ContinuousLegend$get_color_map()}}
\item \href{#method-ContinuousLegend-export}{\code{ContinuousLegend$export()}}
\item \href{#method-ContinuousLegend-clone}{\code{ContinuousLegend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{ContinuousLegend} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$new(min_value, max_value, colors, n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_value}}{\code{numeric} value.}

\item{\code{max_value}}{\code{numeric} value.}

\item{\code{colors}}{\code{character} vector of colors.}

\item{\code{n}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ContinuousLegend} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for creating a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ContinuousLegend} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-get_resample_method"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-get_resample_method}{}}}
\subsection{Method \code{get_resample_method()}}{
Get resample method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$get_resample_method()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-get_color_map"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-get_color_map}{}}}
\subsection{Method \code{get_color_map()}}{
Get a function for mapping values to colors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$get_color_map()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{function} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-export"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-export}{}}}
\subsection{Method \code{export()}}{
Export settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousLegend-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousLegend-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousLegend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
