% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{Dataset}
\alias{Dataset}
\title{Dataset class}
\description{
Dataset class

Dataset class
}
\details{
Definition for the Dataset class.

Note that this method has no effect if the dataset does not have
file path on disk.
}
\seealso{
\code{\link[=new_dataset]{new_dataset()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{\code{character}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:stack]{raster::stack()}} object.}

\item{\code{data}}{\code{NULL}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:stack]{raster::stack()}} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-print}{\code{Dataset$print()}}
\item \href{#method-repr}{\code{Dataset$repr()}}
\item \href{#method-import}{\code{Dataset$import()}}
\item \href{#method-clean}{\code{Dataset$clean()}}
\item \href{#method-get_data}{\code{Dataset$get_data()}}
\item \href{#method-get_variable}{\code{Dataset$get_variable()}}
\item \href{#method-has_variable}{\code{Dataset$has_variable()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(source)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{\code{character}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:stack]{raster::stack()}} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Dataset object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import the dataset into memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
Clean the dataset from memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable"></a>}}
\if{latex}{\out{\hypertarget{method-get_variable}{}}}
\subsection{Method \code{get_variable()}}{
Get a variable from the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_variable(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_variable"></a>}}
\if{latex}{\out{\hypertarget{method-has_variable}{}}}
\subsection{Method \code{has_variable()}}{
Check if the dataset has a variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$has_variable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating if the variable is present or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
