% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{Dataset}
\alias{Dataset}
\title{Dataset class}
\description{
Dataset class

Dataset class
}
\details{
Definition for the Dataset class.

Note that this method has no effect if the dataset does not have
file path on disk.
}
\seealso{
\code{\link[=new_dataset]{new_dataset()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}

\item{\code{spatial_data}}{\code{NULL}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:raster]{raster::raster()}} object.}

\item{\code{attribute_data}}{\code{NULL}, or \code{\link[tibble:tibble]{tibble::tibble()}} object.}

\item{\code{boundary_data}}{\code{NULL}, or \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-print}{\code{Dataset$print()}}
\item \href{#method-repr}{\code{Dataset$repr()}}
\item \href{#method-import}{\code{Dataset$import()}}
\item \href{#method-write}{\code{Dataset$write()}}
\item \href{#method-clean}{\code{Dataset$clean()}}
\item \href{#method-get_spatial_data}{\code{Dataset$get_spatial_data()}}
\item \href{#method-get_attribute_data}{\code{Dataset$get_attribute_data()}}
\item \href{#method-get_boundary_data}{\code{Dataset$get_boundary_data()}}
\item \href{#method-get_crs}{\code{Dataset$get_crs()}}
\item \href{#method-get_planning_unit_indices}{\code{Dataset$get_planning_unit_indices()}}
\item \href{#method-get_planning_unit_areas}{\code{Dataset$get_planning_unit_areas()}}
\item \href{#method-get_index}{\code{Dataset$get_index()}}
\item \href{#method-has_index}{\code{Dataset$has_index()}}
\item \href{#method-max_index}{\code{Dataset$max_index()}}
\item \href{#method-add_index}{\code{Dataset$add_index()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(
  id,
  spatial_path,
  attribute_path,
  boundary_path,
  spatial_data,
  attribute_data,
  boundary_data
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}

\item{\code{spatial_data}}{\code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:raster]{raster::raster()}} object.}

\item{\code{attribute_data}}{\code{\link[tibble:tibble]{tibble::tibble()}} object.}

\item{\code{boundary_data}}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Dataset object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import the data into memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$write(spatial_path, attribute_path, boundary_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spatial_path}}{\code{character} file path.}

\item{\code{attribute_path}}{\code{character} file path.}

\item{\code{boundary_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
Clean the dataset from memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_spatial_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_spatial_data}{}}}
\subsection{Method \code{get_spatial_data()}}{
Get the spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_spatial_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attribute_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_attribute_data}{}}}
\subsection{Method \code{get_attribute_data()}}{
Get the attribute data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_attribute_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_boundary_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_boundary_data}{}}}
\subsection{Method \code{get_boundary_data()}}{
Get the spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_boundary_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_crs"></a>}}
\if{latex}{\out{\hypertarget{method-get_crs}{}}}
\subsection{Method \code{get_crs()}}{
Get the coordinate reference system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_crs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_crs]{sf::st_crs()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_planning_unit_indices"></a>}}
\if{latex}{\out{\hypertarget{method-get_planning_unit_indices}{}}}
\subsection{Method \code{get_planning_unit_indices()}}{
Get planning unit indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_planning_unit_indices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} vector of indices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_planning_unit_areas"></a>}}
\if{latex}{\out{\hypertarget{method-get_planning_unit_areas}{}}}
\subsection{Method \code{get_planning_unit_areas()}}{
Get area values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_planning_unit_areas()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector of values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-get_index}{}}}
\subsection{Method \code{get_index()}}{
Get a data from the dataset at an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_index"></a>}}
\if{latex}{\out{\hypertarget{method-has_index}{}}}
\subsection{Method \code{has_index()}}{
Check if the dataset has an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$has_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical} indicating if data is present or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-max_index"></a>}}
\if{latex}{\out{\hypertarget{method-max_index}{}}}
\subsection{Method \code{max_index()}}{
Maximum index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$max_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} largest index.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_index"></a>}}
\if{latex}{\out{\hypertarget{method-add_index}{}}}
\subsection{Method \code{add_index()}}{
Add data at an index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$add_index(index, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{\code{character} or \code{integer} indicating the field/layer with
the data.}

\item{\code{values}}{\code{numeric} vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
