% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{Dataset}
\alias{Dataset}
\title{Dataset class}
\description{
Dataset class

Dataset class
}
\details{
Definition for the Dataset class.
}
\seealso{
\code{\link[=new_dataset]{new_dataset()}}, \code{\link[=read_dataset]{read_dataset()}}, \code{\link[=read_dataset_with_metadata]{read_dataset_with_metadata()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{\code{character} value.}

\item{\code{subset}}{\code{character} or \code{integer} value indicating the
field/layer within the dataset with the data.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}

\item{\code{legend}}{\code{Legend} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-print}{\code{Dataset$print()}}
\item \href{#method-repr}{\code{Dataset$repr()}}
\item \href{#method-import}{\code{Dataset$import()}}
\item \href{#method-clean}{\code{Dataset$clean()}}
\item \href{#method-get_data}{\code{Dataset$get_data()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(source, subset, total, units, legend)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{\code{character} value.}

\item{\code{subset}}{\code{character} or \code{integer} value.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}

\item{\code{legend}}{\code{Legend} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Dataset object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import the dataset into memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
Clean the dataset from memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
