% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs_functions.R
\name{disable_html_element}
\alias{disable_html_element}
\title{Disable a HTML element}
\usage{
disable_html_element(inputId)
}
\arguments{
\item{inputId}{HTML element id.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Disable a HTML element.
}
