% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Exclude.R
\name{Exclude}
\alias{Exclude}
\title{Exclude class}
\description{
Exclude class

Exclude class
}
\details{
Definition for the Exclude class.
}
\seealso{
\code{\link[=new_exclude]{new_exclude()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} name.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{overlap}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Exclude-new}{\code{Exclude$new()}}
\item \href{#method-Exclude-print}{\code{Exclude$print()}}
\item \href{#method-Exclude-repr}{\code{Exclude$repr()}}
\item \href{#method-Exclude-get_layer_name}{\code{Exclude$get_layer_name()}}
\item \href{#method-Exclude-get_layer_index}{\code{Exclude$get_layer_index()}}
\item \href{#method-Exclude-get_layer_pane}{\code{Exclude$get_layer_pane()}}
\item \href{#method-Exclude-get_id}{\code{Exclude$get_id()}}
\item \href{#method-Exclude-get_visible}{\code{Exclude$get_visible()}}
\item \href{#method-Exclude-get_invisible}{\code{Exclude$get_invisible()}}
\item \href{#method-Exclude-get_loaded}{\code{Exclude$get_loaded()}}
\item \href{#method-Exclude-get_hidden}{\code{Exclude$get_hidden()}}
\item \href{#method-Exclude-get_status}{\code{Exclude$get_status()}}
\item \href{#method-Exclude-get_overlap}{\code{Exclude$get_overlap()}}
\item \href{#method-Exclude-get_data}{\code{Exclude$get_data()}}
\item \href{#method-Exclude-set_new_pane}{\code{Exclude$set_new_pane()}}
\item \href{#method-Exclude-get_setting}{\code{Exclude$get_setting()}}
\item \href{#method-Exclude-set_visible}{\code{Exclude$set_visible()}}
\item \href{#method-Exclude-set_invisible}{\code{Exclude$set_invisible()}}
\item \href{#method-Exclude-set_loaded}{\code{Exclude$set_loaded()}}
\item \href{#method-Exclude-set_status}{\code{Exclude$set_status()}}
\item \href{#method-Exclude-set_setting}{\code{Exclude$set_setting()}}
\item \href{#method-Exclude-get_solution_settings_widget_data}{\code{Exclude$get_solution_settings_widget_data()}}
\item \href{#method-Exclude-get_map_manager_widget_data}{\code{Exclude$get_map_manager_widget_data()}}
\item \href{#method-Exclude-export}{\code{Exclude$export()}}
\item \href{#method-Exclude-render_on_map}{\code{Exclude$render_on_map()}}
\item \href{#method-Exclude-update_on_map}{\code{Exclude$update_on_map()}}
\item \href{#method-Exclude-clone}{\code{Exclude$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-new"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-new}{}}}
\subsection{Method \code{new()}}{
Create a new Exclude object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$new(
  id,
  name,
  variable,
  pane,
  mandatory,
  visible,
  invisible,
  loaded,
  hidden,
  status,
  overlap
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} value.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{overlap}}{\code{character} vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Exclude object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-print"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_layer_name"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_layer_name}{}}}
\subsection{Method \code{get_layer_name()}}{
Get layer names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_layer_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_layer_index"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_layer_index}{}}}
\subsection{Method \code{get_layer_index()}}{
Get layer index values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_layer_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_layer_pane"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_layer_pane}{}}}
\subsection{Method \code{get_layer_pane()}}{
Get layer pane class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_layer_pane()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get exclude identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_invisible}{}}}
\subsection{Method \code{get_invisible()}}{
Get invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} date/time value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_loaded}{}}}
\subsection{Method \code{get_loaded()}}{
Get loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_hidden}{}}}
\subsection{Method \code{get_hidden()}}{
Get hidden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_hidden()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_overlap}{}}}
\subsection{Method \code{get_overlap()}}{
Get overlap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_overlap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_new_pane"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_new_pane}{}}}
\subsection{Method \code{set_new_pane()}}{
Set new pane.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_new_pane(id, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} unique identifier.}

\item{\code{index}}{\code{character} variable index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} or \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_invisible}{}}}
\subsection{Method \code{set_invisible()}}{
Set invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_invisible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} date/time value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_loaded}{}}}
\subsection{Method \code{set_loaded()}}{
Set loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_loaded(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} or `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the exclude in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the exclude in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-export"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-export}{}}}
\subsection{Method \code{export()}}{
Export settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exclude-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Exclude-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exclude$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
