% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-class.R
\name{Feature}
\alias{Feature}
\title{Feature class}
\description{
Feature class

Feature class
}
\details{
Definition for the Feature class.
}
\seealso{
\code{\link[=new_feature]{new_feature()}}, \code{\link[=new_variable]{new_variable()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} unique identifier.}

\item{\code{name}}{\code{character} name.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{goal}}{\code{numeric} value.}

\item{\code{current}}{\code{numeric} value.}

\item{\code{initial_goal}}{\code{numeric} initial goal value.}

\item{\code{min_goal}}{\code{numeric} minimum goal value.}

\item{\code{max_goal}}{\code{numeric} maximum goal value.}

\item{\code{step_goal}}{\code{numeric} step goal value.}

\item{\code{limit_goal}}{\code{numeric} limit goal value.}

\item{\code{current_label}}{\code{character} current label.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Feature$new()}}
\item \href{#method-print}{\code{Feature$print()}}
\item \href{#method-repr}{\code{Feature$repr()}}
\item \href{#method-get_visible}{\code{Feature$get_visible()}}
\item \href{#method-get_status}{\code{Feature$get_status()}}
\item \href{#method-get_goal}{\code{Feature$get_goal()}}
\item \href{#method-get_data}{\code{Feature$get_data()}}
\item \href{#method-set_visible}{\code{Feature$set_visible()}}
\item \href{#method-set_status}{\code{Feature$set_status()}}
\item \href{#method-set_goal}{\code{Feature$set_goal()}}
\item \href{#method-export}{\code{Feature$export()}}
\item \href{#method-clone}{\code{Feature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Feature object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$new(
  id,
  name,
  variable,
  initial_visible,
  initial_status,
  initial_goal,
  min_goal,
  max_goal,
  step_goal,
  limit_goal,
  current,
  current_label,
  icon
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} .}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{initial_goal}}{\code{numeric} value.}

\item{\code{min_goal}}{\code{numeric} value.}

\item{\code{max_goal}}{\code{numeric} value.}

\item{\code{step_goal}}{\code{numeric} value.}

\item{\code{limit_goal}}{\code{numeric} value.}

\item{\code{current}}{\code{numeric} value.}

\item{\code{current_label}}{\code{character} value.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Feature object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_goal"></a>}}
\if{latex}{\out{\hypertarget{method-get_goal}{}}}
\subsection{Method \code{get_goal()}}{
Get goal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$get_goal()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_goal}{}}}
\subsection{Method \code{set_goal()}}{
Set goal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$set_goal(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
