% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_projects.R
\name{find_projects}
\alias{find_projects}
\title{Find projects}
\usage{
find_projects(x)
}
\arguments{
\item{x}{\code{character} directory path.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} object containing the configuration files
and names of projects found in the directory.
}
\description{
Find project files in a directory.
}
\examples{
# find directory with built-in projects
d <- system.file("extdata", "projects", package = "wheretowork")

# list projects in directory
find_projects(d)
}
