% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_horizontalPickerInput.R
\name{horizontalPickerInput}
\alias{horizontalPickerInput}
\title{Horizontal picker input}
\usage{
horizontalPickerInput(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}.}
}
\value{
A select list control that can be added to a user interface
definition.
}
\description{
Create a horizontal \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}.
}
