% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_leaflet_map.R
\name{leaflet_map}
\alias{leaflet_map}
\title{Leaflet map}
\usage{
leaflet_map(sidebar_ids)
}
\arguments{
\item{sidebar_ids}{\code{character} vector containing HTML identifiers for
the left and right sidebars.}
}
\value{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
\description{
Create a leaflet map with customized default settings.
}
