% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{leaflet_map}
\alias{leaflet_map}
\title{Leaflet map}
\usage{
leaflet_map()
}
\value{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
\description{
Create a leaflet map with customized default settings.
}
