% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManager.R
\name{mapManager}
\alias{mapManager}
\title{Map manager}
\usage{
mapManager(mapElementId, items, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{mapElementId}{\code{character} HTML identifer for Leaflet map.
This should be the argument supplied to the \code{outputId} parameter in
the \code{\link[leaflet:map-shiny]{leaflet::leafletOutput()}}.}

\item{items}{\code{list} of objects to display on the map. These objects can
include \link{Weight}, \link{SingleTheme}, \link{MultiTheme} objects.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for managing items on a Leaflet map.
This widget is designed to be used in conjunction with an existing
Leaflet Map within a Shiny web application.
}
\section{Server value}{

The widget sends the following values to the server
(where \code{eId} corresponds to the \code{elementId}):

\describe{

\item{id_run}{\code{character} name to use for a new solution.
This value is updated when the user clicks the button to generate
a new solution.}

\item{id_setting}{\code{list} containng updated weight settings.
Specifically, this object contains the following elements:

\describe{
\item{id}{\code{character} identifier for the theme or weight.}
\item{parameter}{\code{character} name of the updated parameter.
Available options include: \code{"status"}, \code{"factor"}, or \code{"goal"}.}
\item{value}{\code{numeric} new setting value.}
}
}
}
}

