% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTheme-class.R
\name{MultiTheme}
\alias{MultiTheme}
\title{MultiTheme class}
\description{
MultiTheme class

MultiTheme class
}
\details{
Definition for the MultiTheme class.
}
\seealso{
\code{\link[=new_multi_theme]{new_multi_theme()}}.
}
\section{Super class}{
\code{\link[locationmisc:Theme]{locationmisc::Theme}} -> \code{MultiTheme}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group_min_goal}}{\code{numeric} value.}

\item{\code{group_max_goal}}{\code{numeric} value.}

\item{\code{group_initial_goal}}{\code{numeric} value.}

\item{\code{group_limit_goal}}{\code{numeric} value.}

\item{\code{group_step_goal}}{\code{numeric} value.}

\item{\code{group_current_label}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiTheme$new()}}
\item \href{#method-get_solution_settings_widget_data}{\code{MultiTheme$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{MultiTheme$get_map_manager_widget_data()}}
\item \href{#method-clone}{\code{MultiTheme$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_goal">}\href{../../locationmisc/html/Theme.html#method-get_feature_goal}{\code{locationmisc::Theme$get_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_status">}\href{../../locationmisc/html/Theme.html#method-get_feature_status}{\code{locationmisc::Theme$get_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_parameter">}\href{../../locationmisc/html/Theme.html#method-get_parameter}{\code{locationmisc::Theme$get_parameter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="print">}\href{../../locationmisc/html/Theme.html#method-print}{\code{locationmisc::Theme$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="repr">}\href{../../locationmisc/html/Theme.html#method-repr}{\code{locationmisc::Theme$repr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_goal">}\href{../../locationmisc/html/Theme.html#method-set_feature_goal}{\code{locationmisc::Theme$set_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_status">}\href{../../locationmisc/html/Theme.html#method-set_feature_status}{\code{locationmisc::Theme$set_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_parameter">}\href{../../locationmisc/html/Theme.html#method-set_parameter}{\code{locationmisc::Theme$set_parameter()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a MultiTheme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$new(
  id,
  name,
  feature,
  group_min_goal,
  group_max_goal,
  group_initial_goal,
  group_limit_goal,
  group_step_goal,
  group_current_label,
  initial_status,
  round,
  icon
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{group_min_goal}}{\code{numeric} value.}

\item{\code{group_max_goal}}{\code{numeric} value.}

\item{\code{group_initial_goal}}{\code{numeric} value.}

\item{\code{group_limit_goal}}{\code{numeric} value.}

\item{\code{group_step_goal}}{\code{numeric} value.}

\item{\code{group_current_label}}{\code{character} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{round}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MultiTheme object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
