% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTheme-class.R
\name{MultiTheme}
\alias{MultiTheme}
\title{MultiTheme class}
\description{
MultiTheme class

MultiTheme class
}
\details{
Definition for the MultiTheme class.
}
\seealso{
\code{\link[=new_multi_theme]{new_multi_theme()}}.
}
\section{Super class}{
\code{\link[locationmisc:Theme]{locationmisc::Theme}} -> \code{MultiTheme}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group_min_goal}}{\code{numeric} value.}

\item{\code{group_max_goal}}{\code{numeric} value.}

\item{\code{group_initial_goal}}{\code{numeric} value.}

\item{\code{group_limit_goal}}{\code{numeric} value.}

\item{\code{group_step_goal}}{\code{numeric} value.}

\item{\code{group_current_label}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiTheme$new()}}
\item \href{#method-get_new_solution_manager_data}{\code{MultiTheme$get_new_solution_manager_data()}}
\item \href{#method-get_map_manager_data}{\code{MultiTheme$get_map_manager_data()}}
\item \href{#method-clone}{\code{MultiTheme$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_goals">}\href{../../locationmisc/html/Theme.html#method-get_feature_goals}{\code{locationmisc::Theme$get_feature_goals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_statuses">}\href{../../locationmisc/html/Theme.html#method-get_feature_statuses}{\code{locationmisc::Theme$get_feature_statuses()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_goals">}\href{../../locationmisc/html/Theme.html#method-set_feature_goals}{\code{locationmisc::Theme$set_feature_goals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_statuses">}\href{../../locationmisc/html/Theme.html#method-set_feature_statuses}{\code{locationmisc::Theme$set_feature_statuses()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a MultiTheme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$new(
  id,
  name,
  feature,
  group_min_goal,
  group_max_goal,
  group_initial_goal,
  group_limit_goal,
  group_step_goal,
  group_current_label,
  initial_status,
  round,
  icon
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{group_min_goal}}{\code{numeric} value.}

\item{\code{group_max_goal}}{\code{numeric} value.}

\item{\code{group_initial_goal}}{\code{numeric} value.}

\item{\code{group_limit_goal}}{\code{numeric} value.}

\item{\code{group_step_goal}}{\code{numeric} value.}

\item{\code{group_current_label}}{\code{character} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{round}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MultiTheme object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_new_solution_manager_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_new_solution_manager_data}{}}}
\subsection{Method \code{get_new_solution_manager_data()}}{
Get data for displaying the theme in a \link{newSolutionManager} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_new_solution_manager_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with data for
displaying the object in a \code{\link[=newSolutionManager]{newSolutionManager()}} widget.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_data}{}}}
\subsection{Method \code{get_map_manager_data()}}{
Get data for displaying the theme in a \link{mapManager} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$get_map_manager_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with data for displaying the object in a \code{\link[=mapManager]{mapManager()}}
widget.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
