% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{new_dataset}
\alias{new_dataset}
\title{New dataset}
\usage{
new_dataset(source)
}
\arguments{
\item{source}{\code{character}, \code{\link[sf:sf]{sf::st_sf()}}, or \code{\link[raster:stack]{raster::stack()}} object.}
}
\value{
A \link{Dataset} object.
}
\description{
Create a new \link{Dataset} object.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# print object
print(d)

}
