% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-class.R
\name{new_dataset}
\alias{new_dataset}
\title{New dataset}
\usage{
new_dataset(source, subset, total, units, legend)
}
\arguments{
\item{source}{\code{character} file path for the dataset.}

\item{subset}{\code{character} or \code{integer} indicating the field/layer with
data within the dataset.}

\item{total}{\code{numeric} total amount of all values in the underlying data.}

\item{units}{\code{character} units for the values in the underlying data.}

\item{legend}{\code{Legend} object.}
}
\value{
A \link{Dataset} object.
}
\description{
Create a new \link{Dataset} object.
}
\examples{
# create new object
l <- new_dataset(
  source = tempfile(), subset = 1, total = 12, units = "ha",
  legend = new_continuous_legend(1, 100, c("#000000", "#AAAAAA")))

# print object
print(l)

}
