% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Feature.R
\name{new_feature}
\alias{new_feature}
\title{New feature}
\usage{
new_feature(
  name,
  variable,
  visible = TRUE,
  loaded = TRUE,
  hidden = FALSE,
  status = TRUE,
  current = 0,
  goal = 0.2,
  limit_goal = 0,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name of the feature.}

\item{variable}{\link{Variable} object.}

\item{visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}.}

\item{loaded}{\code{logical} The initial loaded value.
This is used to determine if the feature is loaded (or not)
or not the map.
Defaults to \code{FALSE}.}

\item{hidden}{\code{logical} The hidden value.
This is used to determine if the feature is can ever be displayed (or not)
or not the map. Unlike \code{visible}, if this parameter is \code{FALSE} then a
feature can never be viewed on the map.
Defaults to \code{FALSE}.}

\item{status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}.}

\item{current}{\code{numeric} current proportion of values held in existing
conservation areas (e.g. 0.1 = 10\%).}

\item{goal}{\code{numeric} The initial goal for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0.3 (i.e. 30\%).}

\item{limit_goal}{\code{numeric} The minimum goal
(inclusive) that can be selected for the feature.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0 (i.e. 0\%).}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Feature} object.
}
\description{
Create a new \link{Feature} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(dataset = d, index = 1)

# create feature using the variable
f <- new_feature(name = "Intact Alvar", variable = v)

# print object
print(f)
}
