% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureResults-class.R
\name{new_feature_results}
\alias{new_feature_results}
\title{New feature results}
\usage{
new_feature_results(feature, held, id = uuid::UUIDgenerate())
}
\arguments{
\item{feature}{\link{Feature} object.}

\item{held}{\code{numeric} proportion of the feature held by the solution.
(e.g. 0.1 = 10\%).}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{FeatureResults} object.
}
\description{
Create a new \link{FeatureResults} object to store results for a solution.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar", variable = v)

# create a feature results object to store results for the feature
fr <- new_feature_results(f, held = 0.8)

# print object
print(fr)

}
