% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapManager-class.R
\name{new_map_manager}
\alias{new_map_manager}
\title{New map manager}
\usage{
new_map_manager(layers, order = rev(seq_along(layers)))
}
\arguments{
\item{layers}{\code{list} of \link{Theme} and/or \link{Weight} objects.}

\item{order}{\code{numeric} vector containing a value for each element in
\code{layers}. These values indicate the order each layer should
appear on them map. A value of 1 indicates that a layer should appear
beneath every other layer.
Defaults to an integer sequence of numbers based on the number of elements
in \code{layers}.}
}
\value{
A \link{MapManager} object.
}
\description{
Create a new \link{MapManager} object.
}
\examples{
# create dataset
# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# create variables
v1 <- new_variable_from_auto(dataset = d, index = 1)
v2 <- new_variable_from_auto(dataset = d, index = 2)
v3 <- new_variable_from_auto(dataset = d, index = 3)
v4 <- new_variable_from_auto(dataset = d, index = 4)

# create a weight using a variable
w <- new_weight(
  name = "Human Footprint Index", variable = v1,
  initial_factor = 90, initial_status = FALSE, id = "W1")

# create features using variables
f1 <- new_feature(
  name = "Possum", variable = v2,
  initial_goal = 0.2, initial_status = FALSE, current = 0.5, id = "F1")
f2 <- new_feature(
  name = "Forests", variable = v3,
  initial_goal = 0.3, initial_status = FALSE, current = 0.9, id = "F2")
f3 <- new_feature(
  name = "Shrubs", variable = v4,
  initial_goal = 0.6, initial_status = TRUE, current = 0.4, id = "F3")

# create themes using the features
t1 <- new_single_theme("Species", f1, id = "T1")
t2 <- new_multi_theme("Ecoregions", list(f2, f3), id = "T2")

# create a map manager for the themes and weight
mm <- new_map_manager(layers = list(t1, t2, w))

# print object
print(mm)

}
