% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTheme-class.R
\name{new_multi_theme}
\alias{new_multi_theme}
\title{New multi-theme}
\usage{
new_multi_theme(
  name,
  feature,
  initial_status = TRUE,
  round = TRUE,
  icon = "map-marked-alt",
  id = uuid::UUIDgenerate(),
  group_min_goal = 0,
  group_max_goal = 1,
  group_initial_goal = 0.3,
  group_limit_goal = 0.1,
  group_step_goal = 0.01,
  group_current_label = "Current"
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{feature}{\code{list} of \link{Feature} objects.}

\item{initial_status}{\code{logical} The initial status.
This is used to display information on whether the theme is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{round}{\code{logical} should all numbers be rounded to the nearest integer?
Defaults to \code{TRUE}.}

\item{icon}{\code{shiny.tag} Icon to display for the feature
This icon should indicate the type of data that underpin the feature.
Alternatively, the argument can be a \code{character} to automatically
generate a \code{shiny.tag} icon (using \code{\link[shiny:icon]{shiny::icon()}}).
Defaults to \code{"map-marked-alt"}.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}

\item{group_min_goal}{\code{numeric} The minimum goal (inclusive)
shown for the features under the group view.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0 (i.e. 0\%).}

\item{group_max_goal}{\code{numeric} The maximum goal (inclusive)
shown for the features under the group view.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 1 (i.e. 100\%).}

\item{group_initial_goal}{\code{numeric} The initial goal
for the features under the group view.
Note that goal values are specified as proportions, such that a
value of 0.1 corresponds to 10\%.
Defaults to 0.3 (i.e. 30\%).}

\item{group_limit_goal}{\code{numeric}  The minimum increment for setting goals
for the features under the group view.
Note that goal are specified as proportions, such that a
value of 0.01 corresponds to 1\%.
Defaults to 0.01 (i.e. 1\%).}

\item{group_step_goal}{\code{numeric} The minimum increment for setting goals
for the features under the group view.
Note that goal are specified as proportions, such that a
value of 0.01 corresponds to 1\%.
Defaults to 0.01 (i.e. 1\%).}

\item{group_current_label}{\code{character} The display label for the current
level of representation for the features under the group view.
Defaults to \code{Current}.}
}
\value{
A \link{MultiTheme} object.
}
\description{
Create a new \link{MultiTheme} object.
}
\examples{
# create layers
l1 <- new_layer(source = tempfile(), total = 12, units = "ha")
l2 <- new_layer(source = tempfile(), total = 15, units = "ha")
l3 <- new_layer(source = tempfile(), total = 20, units = "ha")

# create features the layers
f1 <- new_feature(name = "Pangolin", layer = l1)
f2 <- new_feature(name = "Panda", layer = l2)
f3 <- new_feature(name = "Palms", layer = l3)

# create a multi-theme using the features
mt <- new_multi_theme(
  name = "Endangered species", feature = list(f1, f2, f3))

# print object
print(mt)

}
