% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MultiTheme.R
\name{new_multi_theme}
\alias{new_multi_theme}
\title{New multi-theme}
\usage{
new_multi_theme(
  name,
  feature,
  feature_order = as.double(rev(seq_along(feature))),
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{feature}{\code{list} of \link{Feature} objects.}

\item{feature_order}{\code{numeric} Relative order for displaying each feature
on a map. Defaults to a reverse sequence of integer values.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{MultiTheme} object.
}
\description{
Create a new \link{MultiTheme} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v1 <- new_variable_from_auto(d, index = 1)
v2 <- new_variable_from_auto(d, index = 2)
v3 <- new_variable_from_auto(d, index = 3)

# create new features
f1 <- new_feature(name = "Pangolin", variable = v1)
f2 <- new_feature(name = "Panda", variable = v2)
f3 <- new_feature(name = "Palms", variable = v3)

# create a theme using the features
mt <- new_multi_theme(
  name = "Endangered species", feature = list(f1, f2, f3)
)

# print object
print(mt)
}
