% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_NullLegend.R
\name{new_null_legend}
\alias{new_null_legend}
\title{New categorical legend}
\usage{
new_null_legend()
}
\arguments{
\item{values}{\code{NULL}}

\item{colors}{\code{NULL}}
}
\value{
A \link{NullLegend} object.
}
\description{
Create a new \link{NullLegend} object.
}
\examples{
# create new object
l <- new_null_legend()

# print object
print(l)
}
