% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Parameter-class.R
\name{new_parameter}
\alias{new_parameter}
\title{New parameter}
\usage{
new_parameter(
  name,
  status = TRUE,
  value = 0,
  min_value = 0,
  max_value = 100,
  step_value = 1,
  hide = FALSE,
  disable = FALSE,
  no_slider = FALSE,
  units = "",
  reference_value = NA_real_,
  reference_units = "",
  tool_tip = "",
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}.}

\item{value}{\code{numeric} initial value.
Defaults to 0.}

\item{min_value}{\code{numeric} minimum value.
Defaults to 0.}

\item{max_value}{\code{numeric} maximum value.}

\item{step_value}{\code{numeric} step value.
Defaults to 1.}

\item{hide}{\code{logical} indicating if the slider should be
hidden when the setting switch is on.
Defaults to \code{FALSE}.}

\item{disable}{\code{logical} indicating if the setting should be disabled.
Defaults to \code{FALSE}.}

\item{no_slider}{\code{logical} indicating if the setting should not have a slider.
Defaults to \code{FALSE}.}

\item{units}{\code{character} units.
Defaults to an empty \code{character} object.}

\item{reference_value}{\code{numeric} reference value.
This parameter is useful if the \code{value} is a relative value,
because the total expected amount (i.e. \code{reference_value * value})
can be reported.
Defaults to \code{NA_real_} indicating that no reference value information
should be reported.}

\item{reference_units}{\code{character} units for the reference value.
Defaults to "".}

\item{tool_tip}{\code{character} description of parameter setting.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Parameter} object.
}
\description{
Create a new \link{Parameter} object.
}
\examples{
# create a new parameter
p <- new_parameter(name = "Spatial clustering")

# print object
print(p)
}
