% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTheme-class.R
\name{new_single_theme}
\alias{new_single_theme}
\title{New single theme}
\usage{
new_single_theme(name, feature, id = uuid::UUIDgenerate())
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{feature}{\link{Feature} object.
For compatibility with \code{\link[=new_multi_theme]{new_multi_theme()}}, a \code{list} object
can also be supplied that contains a single \link{Feature} object.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{SingleTheme} object.
}
\description{
Create a new \link{SingleTheme} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar", variable = v)

# create a theme using the single feature
st <- new_single_theme(name = "Intact Alvar", feature = f)

# print object
print(st)

}
