% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{new_solution}
\alias{new_solution}
\title{New solution}
\usage{
new_solution(
  name,
  variable,
  visible,
  parameters,
  statistics,
  theme_results,
  weight_results,
  include_results,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} name for new solution.}

\item{variable}{\link{Variable} object with the solution.}

\item{visible}{\code{logical} should the solution be visible on a map?}

\item{parameters}{\code{list} of \link{Parameter} objects.}

\item{statistics}{\code{list} of \link{Statistic} objects.}

\item{theme_results}{\code{list} of \link{ThemeResults} objects.}

\item{weight_results}{\code{list} of \link{WeightResults} objects.}

\item{include_results}{\code{list} of \link{IncludeResults} objects.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\description{
Create a new \link{Solution} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create variables
v1 <- new_variable_from_auto(dataset = d, index = 1)
v2 <- new_variable_from_auto(dataset = d, index = 2)

# create features using variables
f1 <- new_feature(
  name = "Possum", variable = v2,
  goal = 0.2, status = FALSE, current = 0.5, id = "F1"
)

# create themes using the features
t1 <- new_theme("Species", f1, id = "T1")

# create a feature results object to store results for the feature
fr1 <- new_feature_results(f1, held = 0.8)

# create a theme results object to store results for the feature
tr1 <- new_theme_results(t1, fr1)

# create parameters
p1 <- new_parameter(name = "Spatial clustering")
p2 <- new_parameter(name = "Optimality gap")

# create a new solution
s <- new_solution(
  name = "solution001",
  variable = v2,
  visible = TRUE,
  parameters = list(p1, p2),
  statistics = list(),
  theme_results = list(tr1),
  weight_results = list(),
  include_results = list(),
  id = "solution1"
)

}
