% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionSettings-class.R
\name{new_solution_settings}
\alias{new_solution_settings}
\title{New solution settings}
\usage{
new_solution_settings(themes, weights)
}
\arguments{
\item{themes}{\code{list} of \link{Theme} objects. D}

\item{weights}{\code{list} of \link{Weight} objects.}
}
\value{
A \link{SolutionSettings} object.
}
\description{
Create a new \link{SolutionSettings} object.
}
\examples{
# create dataset
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")
d <- new_dataset(f)

# create variables
v1 <- new_variable_from_auto(dataset = d, index = 1)
v2 <- new_variable_from_auto(dataset = d, index = 2)
v3 <- new_variable_from_auto(dataset = d, index = 3)
v4 <- new_variable_from_auto(dataset = d, index = 4)

# create a weight using a variable
w <- new_weight(
  name = "Human Footprint Index", variable = v1,
  initial_factor = 90, initial_status = FALSE, id = "W1")

# create features using variables
f1 <- new_feature(
  name = "Possum", variable = v2,
  initial_goal = 0.2, initial_status = FALSE, current = 0.5, id = "F1")
f2 <- new_feature(
  name = "Forests", variable = v3,
  initial_goal = 0.3, initial_status = FALSE, current = 0.9, id = "F2")
f3 <- new_feature(
  name = "Shrubs", variable = v4,
  initial_goal = 0.6, initial_status = TRUE, current = 0.4, id = "F3")

# create themes using the features
t1 <- new_single_theme("Species", f1, id = "T1")
t2 <- new_multi_theme("Ecoregions", list(f2, f3), id = "T2")

# create solution settings using the themes and weight
ss <- new_solution_settings(themes = list(t1, t2), weights = list(w))

# print object
print(ss)

}
