% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic-class.R
\name{new_statistic}
\alias{new_statistic}
\title{New statistic}
\usage{
new_statistic(name, value, units)
}
\arguments{
\item{name}{\code{character} name of statistic.}

\item{value}{\code{numeric} value.}

\item{units}{\code{character} units.}
}
\value{
A \link{Statistic} object.
}
\description{
Create a new \link{Statistic} object.
}
\examples{
# create a statistic
x <- new_statistic(name = "Area", value = 12, units = "ha")

# print object
print(x)

}
