% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme-class.R
\name{new_theme}
\alias{new_theme}
\title{New theme}
\usage{
new_theme(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[=new_single_theme]{new_single_theme()}}
or \code{\link[=new_multi_theme]{new_multi_theme()}}.}
}
\value{
A \link{Theme} object.
}
\description{
Create a new \link{Theme} object. This is a convenience wrapper for
the \code{\link[=new_single_theme]{new_single_theme()}} and \code{\link[=new_multi_theme]{new_multi_theme()}} functions.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar map", variable = v)

# create a theme using the feature
x <- new_theme(name = "Intact Alvar", feature = f)

# print object
print(x)

}
