% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme-class.R
\name{new_theme}
\alias{new_theme}
\title{New theme}
\usage{
new_theme(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[=new_single_theme]{new_single_theme()}}
or \code{\link[=new_multi_theme]{new_multi_theme()}}.}
}
\value{
A \link{Theme} object.
}
\description{
Create a new \link{Theme} object. This is a convenience wrapper for
the \code{\link[=new_single_theme]{new_single_theme()}} and \code{\link[=new_multi_theme]{new_multi_theme()}} functions.
}
\examples{
# create new layer
l <- new_layer(source = tempfile(), total = 12, units = "ha")

# create feature using the layer
f <- new_feature(name = "Intact Alvar Occurrence", layer = l)

# create a theme using the single feature
st <- new_theme(name = "Intact Alvar", feature = f)

# print object
print(st)

}
