% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Theme.R
\name{new_theme}
\alias{new_theme}
\title{New theme}
\usage{
new_theme(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[=new_single_theme]{new_single_theme()}}
or \code{\link[=new_multi_theme]{new_multi_theme()}}.}
}
\value{
A \link{Theme} object.
}
\description{
Create a new \link{Theme} object. This is a convenience wrapper for
the \code{\link[=new_single_theme]{new_single_theme()}} and \code{\link[=new_multi_theme]{new_multi_theme()}} functions.
}
\examples{
# find data path
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar map", variable = v)

# create a theme using the feature
x <- new_theme(name = "Intact Alvar", feature = f)

# print object
print(x)
}
