% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeResults-class.R
\name{new_theme_results}
\alias{new_theme_results}
\title{New theme results}
\usage{
new_theme_results(theme, feature_results, id = uuid::UUIDgenerate())
}
\arguments{
\item{theme}{\link{Theme} object.}

\item{feature_results}{\link{FeatureResults} object or a \code{list} of such objects.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{ThemeResults} object.
}
\description{
Create a new \link{ThemeResults} object to store results for a solution.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar", variable = v)

# create a theme using the single feature
th <- new_single_theme(name = "Intact Alvar", feature = f)

# create a feature results object to store results for the feature
fr <- new_feature_results(f, held = 0.8)

# create a theme results object to store results for the feature
thr <- new_theme_results(th, fr)

# print object
print(thr)

}
