% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ThemeResults.R
\name{new_theme_results}
\alias{new_theme_results}
\title{New theme results}
\usage{
new_theme_results(theme, feature_results, id = uuid::UUIDgenerate())
}
\arguments{
\item{theme}{\link{Theme} object.}

\item{feature_results}{\link{FeatureResults} object or a \code{list} of such objects.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{ThemeResults} object.
}
\description{
Create a new \link{ThemeResults} object to store results for a solution.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
 package = "wheretowork"
)
f3 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
 package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create new feature
f <- new_feature(name = "Intact Alvar", variable = v)

# create a theme using the single feature
th <- new_single_theme(name = "Intact Alvar", feature = f)

# create a feature results object to store results for the feature
fr <- new_feature_results(f, held = 0.8)

# create a theme results object to store results for the feature
thr <- new_theme_results(th, fr)

# print object
print(thr)

}
