% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeResults-class.R
\name{new_theme_results}
\alias{new_theme_results}
\title{New theme results}
\usage{
new_theme_results(name, theme, feature_results, id = uuid::UUIDgenerate())
}
\arguments{
\item{name}{\code{character} name of theme.}

\item{theme}{\link{Theme} object.}

\item{feature_results}{\code{FeatureResults} object or a \code{list} of such objects.}
}
\value{
A \link{ThemeResults} object.
}
\description{
Create a new \link{ThemeResults} object to store results for a solution.
}
\examples{
#TODO

}
