% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable-class.R
\name{new_variable_from_auto}
\alias{new_variable_from_auto}
\title{New variable from automatic calculations}
\usage{
new_variable_from_auto(
  dataset,
  index,
  units = "",
  type = "auto",
  colors = "random"
)
}
\arguments{
\item{dataset}{\code{Dataset} file path for the dataset.}

\item{index}{\code{character} or \code{integer} indicating the field/layer with
the data.}

\item{units}{\code{character} units for the values in the underlying data.}

\item{type}{\code{character} indicating if the data contain
continuous (\code{"continuous"}) or categorical (\code{"categorical"})
numerical values.
Defaults to \code{"auto"} such that data are automatically identified.}

\item{colors}{\code{character} object containing the colors for visualization
(see Details for more information).
Defaults to \code{"random"} such that colors are randomly generated.}
}
\value{
A \link{Variable} object.
}
\description{
Create a new \link{Variable} object by automatically calculating
all metadata from the underlying data.
This function is useful when pre-calculated metadata are not available.
Note this function will take longer to create variables than other
functions because it requires performing geospatial operations.
}
\details{
The argument to \code{colors} can be a vector of different colors
(in hexadecimal format, e.g. \verb{"#112233"), or a single }character\verb{containing the name of a color palette that is used to generate a vector of different colors (see [color_palette()] for more information). The color palette name}"random"` is also available, such that
colors are generated using a randomly selected palette.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# print object
print(v)

}
