% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight-class.R
\name{new_weight}
\alias{new_weight}
\title{New weight}
\usage{
new_weight(
  name,
  dataset,
  initial_status = TRUE,
  initial_factor = 0,
  min_factor = 0,
  max_factor = 100,
  step_factor = 1,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{dataset}{\link{Dataset} object.}

\item{initial_status}{\code{logical} The initial status.
This is used to display information on whether the theme is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{initial_factor}{\code{numeric} initial factor value.
Defaults to 0.}

\item{min_factor}{\code{numeric} minimum factor value.
Defaults to 0.}

\item{max_factor}{\code{numeric} maximum factor value.}

\item{step_factor}{\code{numeric} step factor value.
Defaults to 1.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Weight} object.
}
\description{
Create a new \link{Weight} object.
}
\examples{
# create a new dataset
l <- new_dataset(source = tempfile(), total = 12, units = "ha")

# create a new weight
w <- new_weight(name = "NDVI", dataset = l)

# print object
print(w)

}
