% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight-class.R
\name{new_weight}
\alias{new_weight}
\title{New weight}
\usage{
new_weight(
  name,
  variable,
  initial_visible = TRUE,
  initial_status = TRUE,
  initial_factor = 0,
  min_factor = 0,
  max_factor = 100,
  step_factor = 1,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{variable}{\link{Variable} object.}

\item{initial_visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}}

\item{initial_status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}}

\item{initial_factor}{\code{numeric} initial factor value.
Defaults to 0.}

\item{min_factor}{\code{numeric} minimum factor value.
Defaults to 0.}

\item{max_factor}{\code{numeric} maximum factor value.}

\item{step_factor}{\code{numeric} step factor value.
Defaults to 1.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}
}
\value{
A \link{Weight} object.
}
\description{
Create a new \link{Weight} object.
}
\examples{
# find data path
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")

# create new dataset
d <- new_dataset(f)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new weight
w <- new_weight(name = "NDVI", variable = v)

# print object
print(w)

}
