% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Weight.R
\name{new_weight}
\alias{new_weight}
\title{New weight}
\usage{
new_weight(
  name,
  variable,
  visible = TRUE,
  invisible = NA_real_,
  loaded = TRUE,
  hidden = FALSE,
  status = TRUE,
  current = 0,
  factor = 0,
  id = uuid::UUIDgenerate(),
  pane = paste(uuid::UUIDgenerate(), variable$index, sep = "-")
)
}
\arguments{
\item{name}{\code{character} Name to display.}

\item{variable}{\link{Variable} object.}

\item{visible}{\code{logical} The initial visible value.
This is used to determine if the feature is displayed (or not)
or not the map.
Defaults to \code{TRUE}.}

\item{invisible}{\code{numeric} date/time. A time stamp date given to when a
loaded layer is first turned invisible. This is used to keep track
of loaded invisible layers to offload once the cache threshold has been
reached.
Defaults to \code{NA_real_}.}

\item{loaded}{\code{logical} The initial loaded value.
This is used to determine if the feature has been loaded into the DOM.
Defaults to \code{FALSE}.}

\item{hidden}{\code{logical} The hidden value.
This is used to determine if the feature can ever be displayed.
Unlike \code{visible}, if this parameter is \code{FALSE} then a feature can never
be viewed on the map.
Defaults to \code{FALSE}.}

\item{status}{\code{logical} The initial status value.
This is used to display information on whether the feature is
selected (or not) for subsequent analysis.
Defaults to \code{TRUE}.}

\item{current}{\code{numeric} current proportion of values held in existing
conservation areas (e.g. 0.1 = 10\%).}

\item{factor}{\code{numeric} initial factor value.
Defaults to 0.}

\item{id}{\code{character} unique identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}).}

\item{pane}{\code{character} unique map pane identifier.
Defaults to a random identifier (\code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}) concatenated with
layer index.}
}
\value{
A \link{Weight} object.
}
\description{
Create a new \link{Weight} object.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new weight
w <- new_weight(name = "NDVI", variable = v)

# print object
print(w)
}
