% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSolutionSidebarPane.R
\name{newSolutionSidebarPane}
\alias{newSolutionSidebarPane}
\title{New solution sidebar pane}
\usage{
newSolutionSidebarPane(id, solutionSettingsId = paste0(id, "_settings"))
}
\arguments{
\item{id}{\code{character} identifier for the sidebar pane.}

\item{solutionSettingsId}{\code{character} identifier for the
\code{\link[=solutionSettings]{solutionSettings()}} widget to create within the sidebar pane.
This widget is used to control the settings for new solutions.
Defaults to \code{paste0(id, "_settings")}.}
}
\value{
A \code{shiny.tag} object with the sidebar pane.
}
\description{
Constructs a sidebar pane for generating new solutions.
}
\details{
This is designed to be used as an argument to
\link[leaflet.extras2:sidebar_tabs]{leaflet.extras2::sidebar_tabs} when specifying
the user interface for a Shiny web application.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("newSolutionSidebarPane")
}
}
}
