% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_configuration_file.R
\name{read_configuration_file}
\alias{read_configuration_file}
\title{Read a configuration file}
\usage{
read_configuration_file(x, spatial_path, attribute_path, boundary_path)
}
\arguments{
\item{x}{\code{character} file path for configuration file.}

\item{spatial_path}{\code{character} file path for the spatial data.}

\item{attribute_path}{\code{character} file path for the attribute data.}

\item{boundary_path}{\code{character} file path for the attribute data.}
}
\value{
A \code{list} containing the following elements:
\describe{
\item{name}{A \code{character} value indicating the name.}
\item{dataset}{A \link{Dataset} object.}
\item{themes}{A \code{list} of \link{Theme} objects.}
\item{weights}{A \code{list} of \link{Weight} objects.}
\item{includes}{A \code{list} of \link{Include} objects.}
\item{mode}{A \code{character} value indicating the mode.}
}
}
\description{
Read a configuration file from disk.
The configuration file will contain all the data and settings related
to a set of \link{Theme}, \link{Weight}, and \link{Include} objects.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_data.yaml", package = "locationmisc")
f2 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f4 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# read configuration file
x <- read_configuration_file(f1, f2, f3, f4)

# print results
print(x)

}
