% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_io.R
\name{read_spatial_data}
\alias{read_spatial_data}
\title{Read spatial data}
\usage{
read_spatial_data(x)
}
\arguments{
\item{x}{\code{character} file path.}
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.
}
\description{
Import spatial data from disk.
}
\details{
This function imports data in vector format as a \code{\link[sf:sf]{sf::st_sf()}} object.
It also imports data in raster format as a \code{\link[terra:rast]{terra::rast()}} object.
Note that is will only import the first band in a multi-band raster dataset.
}
\examples{
# read raster data
f1 <- system.file("external/rlogo.grd", package = "raster")
read_spatial_data(f1)

# read vector data
f2 <- system.file("shape/nc.shp", package = "sf")
read_spatial_data(f2)
}
