% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_raster_io.R
\name{readNamedRaster}
\alias{readNamedRaster}
\title{Read a Raster object from disk}
\usage{
readNamedRaster(x, ...)
}
\arguments{
\item{x}{\code{\link[raster:raster]{raster::raster()}} object.}

\item{...}{arguments passed to \link[raster:stack]{raster::stack}.}
}
\value{
A raster object.
}
\description{
Read a \code{\link[raster:raster]{raster::raster()}} object from disk, along with a text file containing
the names of each layer in the object.
}
\details{
This file reads a raster object to disk along with
a text file containing the layer names of the raster.
}
