% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinyjs.R
\name{removeMapPane}
\alias{removeMapPane}
\title{Change file icon and update tool tip}
\usage{
removeMapPane(id)
}
\arguments{
\item{id}{\code{character} CSS selector.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Change file icon and update tool tip
}
