% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{reserve_sizes}
\alias{reserve_sizes}
\title{Reserve sizes}
\usage{
reserve_sizes(x, areas, boundary_matrix)
}
\arguments{
\item{x}{\code{numeric} vector containing the solution values for each planning
unit.}

\item{areas}{\code{numeric} vector containing the area of each planning unit.}

\item{boundary_matrix}{\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} object with boundary length
data for the planning units.}
}
\value{
A \code{numeric} vector containing the total area of each reserve.
}
\description{
Calculate the size of individual reserves in a solution.
}
