% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples}
\usage{
runExample()
}
\description{
Run a 'Shiny' application containing showcasing the widgets implemented
in the \pkg{locationmisc} package.
}
\examples{
if (interactive()) {
  runExample()
}

}
