% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{shapefile_field_names}
\alias{shapefile_field_names}
\title{Shapefile field names}
\usage{
shapefile_field_names(x)
}
\arguments{
\item{x}{\code{character} path.}
}
\value{
\code{character} vector of field names.
}
\description{
Quickly extract the field names in a shapefile.
}
\examples{
shapefile_field_names(system.file("shape/nc.shp", package = "sf")

}
