% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(data, n_single_themes, n_multi_themes, n_weights, lambda = 5)
}
\arguments{
\item{data}{\code{\link[sf:sf]{sf::st_sf()}} or \code{\link[raster:raster]{raster::raster()}} object containing
the spatial data that underpin the planning units.}

\item{n_single_themes}{\code{integer} number of themes with a single feature
to simulate.}

\item{n_multi_themes}{\code{integer} number of themes with a multiple features
to simulate.}

\item{n_weights}{\code{integer} number of weights to simulate.}

\item{lambda}{\code{numeric} lambda parameter for simulating the number
of features for each \link{MultiTheme} object.
This parameter is used as a argument to \code{rpois(n, lambda = lambda)}.
Defaults to 5.}
}
\value{
A \code{list} containing the simulated objects. It has \code{themes} and
\code{weights} elements that contain \link{Theme} and \link{Weight} objects (respectively).
}
\description{
This function simulates \link{Weight}, \link{SingleTheme}, and \link{MultiTheme} objects.
}
\examples{
# import data
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")
d <- raster::raster(f)

# simulate data
x <- simulate_data(
  data  = d, n_single_themes = 3, n_multi_themes = 2, n_weights = 1)

# print results
print(x)

}
\seealso{
\link{new_weight}, \link{new_single_theme}, \link{new_multi_theme}.
}
