% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_solution.R
\name{simulate_solution}
\alias{simulate_solution}
\title{Simulate a new solution}
\usage{
simulate_solution(dataset, themes, weights, includes = list())
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{themes}{\code{list} of \link{Theme} objects.}

\item{weights}{\code{list} of \link{Weight} objects.}

\item{includes}{\code{list} of \link{Include} objects.
Defaults to an empty list such that the solution is not simulated
based on any \link{Include} objects.}
}
\value{
A \link{Solution} object.
}
\description{
This function simulates a \link{Solution} object.
}
