% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_legends.R
\name{simulate_solution_legend}
\alias{simulate_solution_legend}
\title{Simulate solution legend}
\usage{
simulate_solution_legend()
}
\value{
A \link{ManualLegend} object.
}
\description{
This function simulates a legend for a \link{Solution} object.
It is primarily used for testing the package.
}
\examples{
print(simulate_solution_legend())

}
