% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_themes.R
\name{simulate_themes}
\alias{simulate_themes}
\title{Simulate themes}
\usage{
simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{n_single_themes}{\code{integer} number of themes with a single feature
to simulate.}

\item{n_multi_themes}{\code{integer} number of themes with a multiple features
to simulate.}

\item{lambda}{\code{numeric} lambda parameter for simulating the number
of features for each \link{MultiTheme} object.
This parameter is used as a argument to \code{rpois(n, lambda = lambda)}.
Defaults to 5.}
}
\value{
A \code{list} of simulated \link{Theme} objects.
}
\description{
This function simulates \link{SingleTheme} and \link{MultiTheme} objects.
}
\examples{
# import data
f <- system.file("extdata", "sim_raster_data.tif", package = "locationmisc")
d <- new_dataset(raster::raster(f))

# simulate data
x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)

# print results
print(x)

}
\seealso{
\link{new_single_theme}, \link{new_multi_theme}.
}
