% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_weights.R
\name{simulate_weights}
\alias{simulate_weights}
\title{Simulate weights}
\usage{
simulate_weights(dataset, n)
}
\arguments{
\item{dataset}{\link{Dataset} object.}

\item{n}{\code{integer} number of weights to simulate.}
}
\value{
A \code{list} of simulated \link{Weight} objects.
}
\description{
This function simulates \link{Weight} objects.
}
\examples{
# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
x <- simulate_weights(dataset = d, n = 2)

# print results
print(x)

}
\seealso{
\link{new_weight}.
}
