% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTheme-class.R
\name{SingleTheme}
\alias{SingleTheme}
\title{SingleTheme class}
\description{
SingleTheme class

SingleTheme class
}
\details{
Definition for the SingleTheme class.

The relative order is fixed at 1 because each this class
only contains a single feature.
}
\seealso{
\code{\link[=new_single_theme]{new_single_theme()}}.
}
\section{Super class}{
\code{\link[locationmisc:Theme]{locationmisc::Theme}} -> \code{SingleTheme}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SingleTheme$new()}}
\item \href{#method-get_feature_order}{\code{SingleTheme$get_feature_order()}}
\item \href{#method-set_feature_order}{\code{SingleTheme$set_feature_order()}}
\item \href{#method-get_solution_settings_widget_data}{\code{SingleTheme$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{SingleTheme$get_map_manager_widget_data()}}
\item \href{#method-clone}{\code{SingleTheme$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_goal">}\href{../../locationmisc/html/Theme.html#method-get_feature_goal}{\code{locationmisc::Theme$get_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_status">}\href{../../locationmisc/html/Theme.html#method-get_feature_status}{\code{locationmisc::Theme$get_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_feature_visible">}\href{../../locationmisc/html/Theme.html#method-get_feature_visible}{\code{locationmisc::Theme$get_feature_visible()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="get_parameter">}\href{../../locationmisc/html/Theme.html#method-get_parameter}{\code{locationmisc::Theme$get_parameter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="print">}\href{../../locationmisc/html/Theme.html#method-print}{\code{locationmisc::Theme$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="repr">}\href{../../locationmisc/html/Theme.html#method-repr}{\code{locationmisc::Theme$repr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_goal">}\href{../../locationmisc/html/Theme.html#method-set_feature_goal}{\code{locationmisc::Theme$set_feature_goal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_status">}\href{../../locationmisc/html/Theme.html#method-set_feature_status}{\code{locationmisc::Theme$set_feature_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_feature_visible">}\href{../../locationmisc/html/Theme.html#method-set_feature_visible}{\code{locationmisc::Theme$set_feature_visible()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="locationmisc" data-topic="Theme" data-id="set_parameter">}\href{../../locationmisc/html/Theme.html#method-set_parameter}{\code{locationmisc::Theme$set_parameter()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a SingleTheme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$new(id, name, feature, mandatory, round, icon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of a single \link{Feature} object.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{round}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SingleTheme object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_order}{}}}
\subsection{Method \code{get_feature_order()}}{
Get relative order for displaying each feature on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$get_feature_order()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_order}{}}}
\subsection{Method \code{set_feature_order()}}{
Set relative order for displaying features on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$set_feature_order(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \link{solutionSettings} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SingleTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
