% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution-class.R
\name{Solution}
\alias{Solution}
\title{Solution class}
\description{
Solution class

Solution class
}
\details{
Definition for the Solution class.
}
\seealso{
\code{\link[=new_solution]{new_solution()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{name}}{\code{character} name of solution.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{primary_statistics}}{\code{list} of \link{Statistic} objects}

\item{\code{supplementary_statistics}}{\code{list} of \link{Statistic} objects}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Solution$new()}}
\item \href{#method-repr}{\code{Solution$repr()}}
\item \href{#method-print}{\code{Solution$print()}}
\item \href{#method-get_solution_results_widget_data}{\code{Solution$get_solution_results_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Solution$get_map_manager_widget_data()}}
\item \href{#method-clone}{\code{Solution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Solution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$new(
  id,
  name,
  variable,
  primary_statistics,
  supplementary_statistics,
  theme_results,
  weight_results
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{primary_statistics}}{\code{list} of \link{Statistic} objects.}

\item{\code{supplementary_statistics}}{\code{list} of \link{Statistic} objects.}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Solution object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_results_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_results_widget_data}{}}}
\subsection{Method \code{get_solution_results_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_solution_results_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
