% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution-class.R
\name{Solution}
\alias{Solution}
\title{Solution class}
\description{
Solution class

Solution class
}
\details{
Definition for the Solution class.
}
\seealso{
\code{\link[=new_solution]{new_solution()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{name}}{\code{character} name of solution.}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{statistics}}{\code{list} of \link{Statistic} objects}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Solution$new()}}
\item \href{#method-repr}{\code{Solution$repr()}}
\item \href{#method-print}{\code{Solution$print()}}
\item \href{#method-get_visible}{\code{Solution$get_visible()}}
\item \href{#method-get_parameter}{\code{Solution$get_parameter()}}
\item \href{#method-set_visible}{\code{Solution$set_visible()}}
\item \href{#method-set_parameter}{\code{Solution$set_parameter()}}
\item \href{#method-get_solution_results_widget_data}{\code{Solution$get_solution_results_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Solution$get_map_manager_widget_data()}}
\item \href{#method-render_on_map}{\code{Solution$render_on_map()}}
\item \href{#method-update_on_map}{\code{Solution$update_on_map()}}
\item \href{#method-clone}{\code{Solution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Solution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$new(
  id,
  name,
  variable,
  initial_visible,
  statistics,
  theme_results,
  weight_results
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{statistics}}{\code{list} of \link{Statistic} objects.}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Solution object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_parameter(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_parameter(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_results_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_results_widget_data}{}}}
\subsection{Method \code{get_solution_results_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_solution_results_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
