% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{Solution}
\alias{Solution}
\title{Solution class}
\description{
Solution class

Solution class
}
\details{
Definition for the Solution class.
}
\seealso{
\code{\link[=new_solution]{new_solution()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{name}}{\code{character} name of solution.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} name.}

\item{\code{parameters}}{\code{list} of \link{Parameter} objects}

\item{\code{statistics}}{\code{list} of \link{Statistic} objects}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}

\item{\code{include_results}}{\code{list} of \link{IncludeResults} objects.}

\item{\code{exclude_results}}{\code{list} of \link{ExcludeResults} objects.}

\item{\code{hidden}}{\code{logical} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Solution-new}{\code{Solution$new()}}
\item \href{#method-Solution-repr}{\code{Solution$repr()}}
\item \href{#method-Solution-print}{\code{Solution$print()}}
\item \href{#method-Solution-get_layer_name}{\code{Solution$get_layer_name()}}
\item \href{#method-Solution-get_layer_index}{\code{Solution$get_layer_index()}}
\item \href{#method-Solution-get_layer_pane}{\code{Solution$get_layer_pane()}}
\item \href{#method-Solution-get_id}{\code{Solution$get_id()}}
\item \href{#method-Solution-get_visible}{\code{Solution$get_visible()}}
\item \href{#method-Solution-get_invisible}{\code{Solution$get_invisible()}}
\item \href{#method-Solution-get_loaded}{\code{Solution$get_loaded()}}
\item \href{#method-Solution-get_hidden}{\code{Solution$get_hidden()}}
\item \href{#method-Solution-get_setting}{\code{Solution$get_setting()}}
\item \href{#method-Solution-set_new_pane}{\code{Solution$set_new_pane()}}
\item \href{#method-Solution-set_visible}{\code{Solution$set_visible()}}
\item \href{#method-Solution-set_invisible}{\code{Solution$set_invisible()}}
\item \href{#method-Solution-set_loaded}{\code{Solution$set_loaded()}}
\item \href{#method-Solution-get_summary_results_data}{\code{Solution$get_summary_results_data()}}
\item \href{#method-Solution-get_theme_results_data}{\code{Solution$get_theme_results_data()}}
\item \href{#method-Solution-get_weight_results_data}{\code{Solution$get_weight_results_data()}}
\item \href{#method-Solution-get_include_results_data}{\code{Solution$get_include_results_data()}}
\item \href{#method-Solution-get_exclude_results_data}{\code{Solution$get_exclude_results_data()}}
\item \href{#method-Solution-render_summary_results}{\code{Solution$render_summary_results()}}
\item \href{#method-Solution-render_theme_results}{\code{Solution$render_theme_results()}}
\item \href{#method-Solution-render_weight_results}{\code{Solution$render_weight_results()}}
\item \href{#method-Solution-render_include_results}{\code{Solution$render_include_results()}}
\item \href{#method-Solution-render_exclude_results}{\code{Solution$render_exclude_results()}}
\item \href{#method-Solution-set_setting}{\code{Solution$set_setting()}}
\item \href{#method-Solution-get_solution_results_widget_data}{\code{Solution$get_solution_results_widget_data()}}
\item \href{#method-Solution-get_map_manager_widget_data}{\code{Solution$get_map_manager_widget_data()}}
\item \href{#method-Solution-render_on_map}{\code{Solution$render_on_map()}}
\item \href{#method-Solution-update_on_map}{\code{Solution$update_on_map()}}
\item \href{#method-Solution-clone}{\code{Solution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-new"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-new}{}}}
\subsection{Method \code{new()}}{
Create a Solution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$new(
  id,
  name,
  variable,
  pane,
  visible,
  invisible,
  loaded,
  statistics,
  parameters,
  theme_results,
  weight_results,
  include_results,
  exclude_results,
  hidden
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{statistics}}{\code{list} of \link{Statistic} objects.}

\item{\code{parameters}}{\code{list} of \link{Statistic} objects.}

\item{\code{theme_results}}{\code{list} of \link{ThemeResults} objects.}

\item{\code{weight_results}}{\code{list} of \link{WeightResults} objects.}

\item{\code{include_results}}{\code{list} of \link{IncludeResults} objects.}

\item{\code{exclude_results}}{\code{list} of \link{ExcludeResults} objects.}

\item{\code{hidden}}{\code{logical} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Solution object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-print"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_layer_name"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_layer_name}{}}}
\subsection{Method \code{get_layer_name()}}{
Get layer names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_layer_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_layer_index"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_layer_index}{}}}
\subsection{Method \code{get_layer_index()}}{
Get layer index values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_layer_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_layer_pane"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_layer_pane}{}}}
\subsection{Method \code{get_layer_pane()}}{
Get layer pane class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_layer_pane()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get solution identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_invisible}{}}}
\subsection{Method \code{get_invisible()}}{
Get invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} date/time value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_loaded}{}}}
\subsection{Method \code{get_loaded()}}{
Get loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_hidden}{}}}
\subsection{Method \code{get_hidden()}}{
Get hidden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_hidden()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-set_new_pane"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-set_new_pane}{}}}
\subsection{Method \code{set_new_pane()}}{
Set new pane.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_new_pane(id, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} unique identifier.}

\item{\code{index}}{\code{character} variable index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-set_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-set_invisible}{}}}
\subsection{Method \code{set_invisible()}}{
Set invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_invisible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} date/time value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-set_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-set_loaded}{}}}
\subsection{Method \code{set_loaded()}}{
Set loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_loaded(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_summary_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_summary_results_data}{}}}
\subsection{Method \code{get_summary_results_data()}}{
Get summary results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_summary_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_theme_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_theme_results_data}{}}}
\subsection{Method \code{get_theme_results_data()}}{
Get theme results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_theme_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_weight_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_weight_results_data}{}}}
\subsection{Method \code{get_weight_results_data()}}{
Get weight results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_weight_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_include_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_include_results_data}{}}}
\subsection{Method \code{get_include_results_data()}}{
Get include results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_include_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_exclude_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_exclude_results_data}{}}}
\subsection{Method \code{get_exclude_results_data()}}{
Get exclude results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_exclude_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_summary_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_summary_results}{}}}
\subsection{Method \code{render_summary_results()}}{
Render summary results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_summary_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_theme_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_theme_results}{}}}
\subsection{Method \code{render_theme_results()}}{
Render theme results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_theme_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_weight_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_weight_results}{}}}
\subsection{Method \code{render_weight_results()}}{
Render weight results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_weight_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_include_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_include_results}{}}}
\subsection{Method \code{render_include_results()}}{
Render include results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_include_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_exclude_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_exclude_results}{}}}
\subsection{Method \code{render_exclude_results()}}{
Render exclude results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_exclude_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_solution_results_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_solution_results_widget_data}{}}}
\subsection{Method \code{get_solution_results_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionResults]{solutionResults()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_solution_results_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
