% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionResultsModal.R
\name{solutionResultsModal}
\alias{solutionResultsModal}
\title{Solution results modal}
\usage{
solutionResultsModal(id)
}
\arguments{
\item{id}{\code{character} identifier for the sidebar pane.}
}
\value{
A \code{shiny.tag} object with the modal
}
\description{
Constructs a modal for displaying solution results.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("solutionResultsModal")
}
}

}
