% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionSettings-class.R
\name{SolutionSettings}
\alias{SolutionSettings}
\title{Solution settings class}
\description{
Solution settings class

Solution settings class
}
\details{
Definition for the SolutionSettings class.

The argument to \code{value} should be a \code{list} with the following elements:
\describe{
\item{id}{\code{character} identifier for theme or weight.}
\item{parameter}{\code{character} name of parameter.
Available options are: \code{"status"}, \code{"factor"}, or \code{"goal"}.}
\item{type}{\code{character} indicating if the parameter is for a theme or
weight.
Available options are: \code{"theme"} or \code{"weight"}.}
}

The argument to \code{value} should be a \code{list} with the following elements:
\describe{
\item{id}{\code{character} identifier for theme or weight.}
\item{parameter}{\code{character} name of parameter.
Available options are: \code{"status"}, \code{"factor"}, or \code{"goal"}.}
\item{value}{\code{numeric} or \code{logical} value for new parameter.}
\item{type}{\code{character} indicating if the parameter is for a theme or
weight.
Available options are: \code{"theme"} or \code{"weight"}.}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{theme_ids}}{\code{character} vector of identifiers for the themes.}

\item{\code{weight_ids}}{\code{character} vector of identifiers for the weights.}

\item{\code{themes}}{\code{list} of \link{Theme} objects.}

\item{\code{weights}}{\code{list} of \link{Weight} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolutionSettings$new()}}
\item \href{#method-print}{\code{SolutionSettings$print()}}
\item \href{#method-repr}{\code{SolutionSettings$repr()}}
\item \href{#method-get_theme}{\code{SolutionSettings$get_theme()}}
\item \href{#method-get_weight}{\code{SolutionSettings$get_weight()}}
\item \href{#method-get_parameter}{\code{SolutionSettings$get_parameter()}}
\item \href{#method-set_parameter}{\code{SolutionSettings$set_parameter()}}
\item \href{#method-get_widget_data}{\code{SolutionSettings$get_widget_data()}}
\item \href{#method-get_prioritization_data}{\code{SolutionSettings$get_prioritization_data()}}
\item \href{#method-clone}{\code{SolutionSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a SolutionSettings object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$new(themes, weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{themes}}{\code{list} of \link{Theme} objects.}

\item{\code{weights}}{\code{list} of \link{Weight} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SolutionSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme}{}}}
\subsection{Method \code{get_theme()}}{
Get a theme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_theme(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} theme identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Theme} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight}{}}}
\subsection{Method \code{get_weight()}}{
Get a weight.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_weight(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} weight identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Weight} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get feature (rij) matrix.


Get a parameter for a weight or theme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_parameter(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} weight identifier.}

\item{\code{value}}{\code{list} with new parameter information (see Details section)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Weight} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set a parameter for a weight or theme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$set_parameter(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{list} with new parameter information (see Details section)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prioritization_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_prioritization_data}{}}}
\subsection{Method \code{get_prioritization_data()}}{
Get data for generating a prioritization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_prioritization_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
