% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionSettings.R
\name{solutionSettings}
\alias{solutionSettings}
\title{Solution settings widget}
\usage{
solutionSettings(x, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{x}{\link{SolutionSettings} object.}

\item{width}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{height}{\code{character} width of the displayed widget.
Defaults to \code{NULL} such that the widget is automatically sized.}

\item{elementId}{\code{character} HTML identifier for the widget.
Defaults to \code{NULL}.}
}
\description{
Constructs a widget for maanging the settings for generating solutions.
}
\section{Server value}{

The widget sends the following values to the server
(where \code{elementId} corresponds to the \code{elementId}):

\describe{

\item{id}{\code{character} identifier for the theme or weight.}

\item{parameter}{\code{character} name of the updated parameter.
Available options include: \code{"status"}, \code{"factor"}, or \code{"goal"}.}

\item{value}{new \code{numeric} or \code{logical} values.}

\item{type}{\code{character} indicating if the updated parameter corresponds
to a \code{theme} or \code{weight}.}

}
}

\examples{
# TODO.

}
