% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SolutionSettings.R
\name{SolutionSettings}
\alias{SolutionSettings}
\title{Solution settings class}
\description{
Solution settings class

Solution settings class
}
\details{
Definition for the SolutionSettings class.

The argument to \code{value} should be a \code{list} with the following elements:
\describe{
\item{id}{\code{character} identifier for theme, weight, or include.}
\item{setting}{\code{character} name of parameter.
Available options are: \code{"status"}, \code{"factor"}, \code{"value"}, or \code{"goal"}.}
\item{type}{\code{character} indicating the type of setting.
Available options are: \code{"theme"}, \code{"weight"}, \code{"include"},
\code{"parameter"}.}
}

The argument to \code{value} should be a \code{list} with the following elements:
\describe{
\item{id}{\code{character} identifier for theme, weight, or include.}
\item{setting}{\code{character} name of parameter.
Available options are: \code{"status"}, \code{"factor"}, \code{"value"}, or \code{"goal"}.}
\item{value}{\code{numeric} or \code{logical} value for new setting.}
\item{type}{\code{character} indicating the type of setting.
Available options are: \code{"theme"}, \code{"weight"}, \code{"include"},
\code{"parameter"}.}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{theme_ids}}{\code{character} vector of identifiers for the themes.}

\item{\code{weight_ids}}{\code{character} vector of identifiers for the weights.}

\item{\code{include_ids}}{\code{character} vector of identifiers for the includes.}

\item{\code{parameter_ids}}{\code{character} vector of identifiers for the
parameters.}

\item{\code{themes}}{\code{list} of \link{Theme} objects.}

\item{\code{weights}}{\code{list} of \link{Weight} objects.}

\item{\code{includes}}{\code{list} of \link{Include} objects.}

\item{\code{parameters}}{\code{list} of \link{Parameter} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolutionSettings$new()}}
\item \href{#method-print}{\code{SolutionSettings$print()}}
\item \href{#method-repr}{\code{SolutionSettings$repr()}}
\item \href{#method-get_theme}{\code{SolutionSettings$get_theme()}}
\item \href{#method-get_weight}{\code{SolutionSettings$get_weight()}}
\item \href{#method-get_include}{\code{SolutionSettings$get_include()}}
\item \href{#method-get_parameter}{\code{SolutionSettings$get_parameter()}}
\item \href{#method-get_setting}{\code{SolutionSettings$get_setting()}}
\item \href{#method-set_setting}{\code{SolutionSettings$set_setting()}}
\item \href{#method-get_widget_data}{\code{SolutionSettings$get_widget_data()}}
\item \href{#method-get_theme_settings}{\code{SolutionSettings$get_theme_settings()}}
\item \href{#method-get_weight_settings}{\code{SolutionSettings$get_weight_settings()}}
\item \href{#method-get_include_settings}{\code{SolutionSettings$get_include_settings()}}
\item \href{#method-get_parameter_settings}{\code{SolutionSettings$get_parameter_settings()}}
\item \href{#method-get_theme_data}{\code{SolutionSettings$get_theme_data()}}
\item \href{#method-get_weight_data}{\code{SolutionSettings$get_weight_data()}}
\item \href{#method-get_include_data}{\code{SolutionSettings$get_include_data()}}
\item \href{#method-update_current_held}{\code{SolutionSettings$update_current_held()}}
\item \href{#method-clone}{\code{SolutionSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a SolutionSettings object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$new(themes, weights, includes, parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{themes}}{\code{list} of \link{Theme} objects.}

\item{\code{weights}}{\code{list} of \link{Weight} objects.}

\item{\code{includes}}{\code{list} of \link{Include} objects.}

\item{\code{parameters}}{\code{list} of \link{Parameter} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SolutionSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme}{}}}
\subsection{Method \code{get_theme()}}{
Get a theme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_theme(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} theme identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Theme} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight}{}}}
\subsection{Method \code{get_weight()}}{
Get a weight.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_weight(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} weight identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Weight} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_include"></a>}}
\if{latex}{\out{\hypertarget{method-get_include}{}}}
\subsection{Method \code{get_include()}}{
Get an include.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_include(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} weight identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Include} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get an parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_parameter(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{character} weight identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Parameter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get a setting for a weight, theme, include, or parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_setting(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{list} with new parameter information (see Details section)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set a setting for a weight or theme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$set_setting(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{list} with new setting information (see Details section)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme_settings"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme_settings}{}}}
\subsection{Method \code{get_theme_settings()}}{
Get theme settings for generating a prioritization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_theme_settings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight_settings"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight_settings}{}}}
\subsection{Method \code{get_weight_settings()}}{
Get weight settings for generating a prioritization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_weight_settings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_include_settings"></a>}}
\if{latex}{\out{\hypertarget{method-get_include_settings}{}}}
\subsection{Method \code{get_include_settings()}}{
Get include settings for generating a prioritization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_include_settings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter_settings"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter_settings}{}}}
\subsection{Method \code{get_parameter_settings()}}{
Get parameter settings for generating a prioritization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_parameter_settings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[tibble:tibble]{tibble::tibble()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme_data}{}}}
\subsection{Method \code{get_theme_data()}}{
Get theme matrix data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_theme_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight_data}{}}}
\subsection{Method \code{get_weight_data()}}{
Get weight matrix data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_weight_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_include_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_include_data}{}}}
\subsection{Method \code{get_include_data()}}{
Get includes matrix data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$get_include_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} with data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_current_held"></a>}}
\if{latex}{\out{\hypertarget{method-update_current_held}{}}}
\subsection{Method \code{update_current_held()}}{
Update the current amount held for each themes and weights automatically
based on the include statuses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$update_current_held(
  theme_data = self$get_theme_data(),
  weight_data = self$get_weight_data(),
  include_data = self$get_include_data()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme_data}}{\verb{[Matrix::sparseMatrix()] with theme data. Defaults to }self$get_theme_data()`.}

\item{\code{weight_data}}{\verb{[Matrix::sparseMatrix()] with weight data. Defaults to }self$get_weight_data()`.}

\item{\code{include_data}}{\verb{[Matrix::sparseMatrix()] with include data. Defaults to }self$get_include_data()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolutionSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
