% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_data_statistics.R
\name{spatial_data_statistics}
\alias{spatial_data_statistics}
\alias{spatial_data_statistics.sf}
\alias{spatial_data_statistics.SpatRaster}
\alias{spatial_data_statistics.data.frame}
\title{Spatial data statistics}
\usage{
spatial_data_statistics(x, type, index = 1)

\method{spatial_data_statistics}{sf}(x, type, index = 1)

\method{spatial_data_statistics}{SpatRaster}(x, type, index = 1)

\method{spatial_data_statistics}{data.frame}(x, type, index = 1)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} or a combined \link[terra:rast]{terra::rast} or \link{data.frame} dataset object.}

\item{type}{\code{character} indicating whether the dataset contains
continuous (i.e. \code{"continuous"}) or categorical (i.e. \code{"categorical"})
data.}

\item{index}{\code{integer} or \code{character} value indicating the
field or layer for which to calculate statistics.
Defaults to 1, such that the first field/layer is used to calculate
statistics.}
}
\value{
A \code{list} containing statistics for the data.
}
\description{
Calculate statistics for a spatial dataset.
}
