% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_type.R
\name{spatial_data_type}
\alias{spatial_data_type}
\alias{spatial_data_type.sf}
\alias{spatial_data_type.Raster}
\title{Spatial data type}
\usage{
spatial_data_type(x, subset = 1)

\method{spatial_data_type}{sf}(x, subset = 1)

\method{spatial_data_type}{Raster}(x, subset = 1, max_sample = 1000)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} or \code{\link[raster:stack]{raster::stack()}} dataset object.}

\item{subset}{\code{integer} or \code{character} value indicating the
field or layer for which to calculate statistics.
Defaults to 1, such that the first field/layer is used to calculate
statistics.}

\item{max_sample}{\code{integer} maximum number of cells in a raster stack
to sample when checking if the data are continuous or categorical.
Defaults to 1000.}
}
\value{
A \code{character} indicating if the data are \code{"continuous"} or
\code{"categorical"}.
}
\description{
Identify if a spatial dataset has continuous or categorical data.
}
