% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme-class.R
\name{Theme}
\alias{Theme}
\title{Theme class}
\description{
Theme class

Theme class
}
\details{
Definition for the Theme class. This is a parent class that contains
shared fields and methods for the \link{SingleTheme} and \link{MultiTheme} classes.

This method intentionally throws an error.
}
\seealso{
\code{\link[=new_single_theme]{new_single_theme()}}, \code{\link[=new_multi_theme]{new_multi_theme()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Theme$new()}}
\item \href{#method-print}{\code{Theme$print()}}
\item \href{#method-repr}{\code{Theme$repr()}}
\item \href{#method-get_feature_id}{\code{Theme$get_feature_id()}}
\item \href{#method-get_feature_name}{\code{Theme$get_feature_name()}}
\item \href{#method-get_feature_current}{\code{Theme$get_feature_current()}}
\item \href{#method-get_feature_total}{\code{Theme$get_feature_total()}}
\item \href{#method-get_feature_visible}{\code{Theme$get_feature_visible()}}
\item \href{#method-get_feature_status}{\code{Theme$get_feature_status()}}
\item \href{#method-get_feature_goal}{\code{Theme$get_feature_goal()}}
\item \href{#method-get_setting}{\code{Theme$get_setting()}}
\item \href{#method-set_feature_visible}{\code{Theme$set_feature_visible()}}
\item \href{#method-set_feature_status}{\code{Theme$set_feature_status()}}
\item \href{#method-set_feature_goal}{\code{Theme$set_feature_goal()}}
\item \href{#method-set_setting}{\code{Theme$set_setting()}}
\item \href{#method-export}{\code{Theme$export()}}
\item \href{#method-render_on_map}{\code{Theme$render_on_map()}}
\item \href{#method-update_on_map}{\code{Theme$update_on_map()}}
\item \href{#method-clone}{\code{Theme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Theme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_id}{}}}
\subsection{Method \code{get_feature_id()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with identifier(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_name}{}}}
\subsection{Method \code{get_feature_name()}}{
Get feature names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with identifier(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_current"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_current}{}}}
\subsection{Method \code{get_feature_current()}}{
Get feature current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_current()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_total"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_total}{}}}
\subsection{Method \code{get_feature_total()}}{
Get feature current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_total()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_visible}{}}}
\subsection{Method \code{get_feature_visible()}}{
Get feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_status}{}}}
\subsection{Method \code{get_feature_status()}}{
Get feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_goal}{}}}
\subsection{Method \code{get_feature_goal()}}{
Get feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_goal()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with goal value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, or \code{"feature_order"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_visible}{}}}
\subsection{Method \code{set_feature_visible()}}{
Set feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_status}{}}}
\subsection{Method \code{set_feature_status()}}{
Set feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goal}{}}}
\subsection{Method \code{set_feature_goal()}}{
Set feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_goal(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector containing a value for each feature.
A \code{list} of \code{numeric} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, \code{"feature_order"}.}

\item{\code{value}}{vector containing a value for each feature.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
