% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme-class.R
\name{Theme}
\alias{Theme}
\title{Theme class}
\description{
Theme class

Theme class
}
\details{
Definition for the Theme class. This is a parent class that contains
shared fields and methods for the \link{SingleTheme} and \link{MultiTheme} classes.

This method intentionally throws an error.
}
\seealso{
\code{\link[=new_single_theme]{new_single_theme()}}, \code{\link[=new_multi_theme]{new_multi_theme()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{mandatory}}{\code{logical} value.}

\item{\code{round}}{\code{logical} value.}

\item{\code{icon}}{\code{shiny.tag} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Theme$new()}}
\item \href{#method-print}{\code{Theme$print()}}
\item \href{#method-repr}{\code{Theme$repr()}}
\item \href{#method-get_feature_visible}{\code{Theme$get_feature_visible()}}
\item \href{#method-get_feature_status}{\code{Theme$get_feature_status()}}
\item \href{#method-get_feature_goal}{\code{Theme$get_feature_goal()}}
\item \href{#method-get_parameter}{\code{Theme$get_parameter()}}
\item \href{#method-set_feature_visible}{\code{Theme$set_feature_visible()}}
\item \href{#method-set_feature_status}{\code{Theme$set_feature_status()}}
\item \href{#method-set_feature_goal}{\code{Theme$set_feature_goal()}}
\item \href{#method-set_parameter}{\code{Theme$set_parameter()}}
\item \href{#method-clone}{\code{Theme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Theme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_visible}{}}}
\subsection{Method \code{get_feature_visible()}}{
Get feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_status}{}}}
\subsection{Method \code{get_feature_status()}}{
Get feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_goal}{}}}
\subsection{Method \code{get_feature_goal()}}{
Get feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_goal()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with goal value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_parameter(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, or \code{"feature_order"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_visible}{}}}
\subsection{Method \code{set_feature_visible()}}{
Set feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_status}{}}}
\subsection{Method \code{set_feature_status()}}{
Set feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goal}{}}}
\subsection{Method \code{set_feature_goal()}}{
Set feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_goal(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector containing a value for each feature.
A \code{list} of \code{numeric} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_parameter(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, \code{"feature_order"}.}

\item{\code{value}}{vector containing a value for each feature.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
