% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Theme.R
\name{Theme}
\alias{Theme}
\title{Theme class}
\description{
Theme class

Theme class
}
\details{
Definition for the Theme class.
}
\seealso{
\code{\link[=new_theme]{new_theme()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{feature_order}}{\code{numeric} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Theme$new()}}
\item \href{#method-print}{\code{Theme$print()}}
\item \href{#method-repr}{\code{Theme$repr()}}
\item \href{#method-get_layer_name}{\code{Theme$get_layer_name()}}
\item \href{#method-get_layer_id}{\code{Theme$get_layer_id()}}
\item \href{#method-get_layer_pane}{\code{Theme$get_layer_pane()}}
\item \href{#method-get_layer_index}{\code{Theme$get_layer_index()}}
\item \href{#method-get_visible}{\code{Theme$get_visible()}}
\item \href{#method-get_invisible}{\code{Theme$get_invisible()}}
\item \href{#method-get_loaded}{\code{Theme$get_loaded()}}
\item \href{#method-get_feature_id}{\code{Theme$get_feature_id()}}
\item \href{#method-get_feature_name}{\code{Theme$get_feature_name()}}
\item \href{#method-get_feature_current}{\code{Theme$get_feature_current()}}
\item \href{#method-get_feature_limit}{\code{Theme$get_feature_limit()}}
\item \href{#method-get_feature_total}{\code{Theme$get_feature_total()}}
\item \href{#method-get_feature_visible}{\code{Theme$get_feature_visible()}}
\item \href{#method-get_feature_invisible}{\code{Theme$get_feature_invisible()}}
\item \href{#method-get_feature_loaded}{\code{Theme$get_feature_loaded()}}
\item \href{#method-get_feature_hidden}{\code{Theme$get_feature_hidden()}}
\item \href{#method-get_feature_status}{\code{Theme$get_feature_status()}}
\item \href{#method-get_feature_goal}{\code{Theme$get_feature_goal()}}
\item \href{#method-get_feature_order}{\code{Theme$get_feature_order()}}
\item \href{#method-set_feature_order}{\code{Theme$set_feature_order()}}
\item \href{#method-set_loaded}{\code{Theme$set_loaded()}}
\item \href{#method-get_setting}{\code{Theme$get_setting()}}
\item \href{#method-set_visible}{\code{Theme$set_visible()}}
\item \href{#method-set_feature_visible}{\code{Theme$set_feature_visible()}}
\item \href{#method-set_feature_status}{\code{Theme$set_feature_status()}}
\item \href{#method-set_feature_goal}{\code{Theme$set_feature_goal()}}
\item \href{#method-set_feature_current}{\code{Theme$set_feature_current()}}
\item \href{#method-set_setting}{\code{Theme$set_setting()}}
\item \href{#method-get_solution_settings_widget_data}{\code{Theme$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Theme$get_map_manager_widget_data()}}
\item \href{#method-render_on_map}{\code{Theme$render_on_map()}}
\item \href{#method-update_on_map}{\code{Theme$update_on_map()}}
\item \href{#method-export}{\code{Theme$export()}}
\item \href{#method-clone}{\code{Theme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Theme object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(id, name, feature, feature_order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{feature}}{\code{list} of \link{Feature} objects.}

\item{\code{feature_order}}{\code{numeric} vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Theme object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_name}{}}}
\subsection{Method \code{get_layer_name()}}{
Get layer names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_layer_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_id}{}}}
\subsection{Method \code{get_layer_id()}}{
Get layer id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_layer_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_pane"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_pane}{}}}
\subsection{Method \code{get_layer_pane()}}{
Get layer map pane classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_layer_pane()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_index"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_index}{}}}
\subsection{Method \code{get_layer_index()}}{
Get layer index values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_layer_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible value for all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-get_invisible}{}}}
\subsection{Method \code{get_invisible()}}{
Get invisible value for all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector of date/times.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-get_loaded}{}}}
\subsection{Method \code{get_loaded()}}{
Get loaded value for all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_id}{}}}
\subsection{Method \code{get_feature_id()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with identifier(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_name}{}}}
\subsection{Method \code{get_feature_name()}}{
Get feature names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with identifier(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_current"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_current}{}}}
\subsection{Method \code{get_feature_current()}}{
Get feature current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_current()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_limit"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_limit}{}}}
\subsection{Method \code{get_feature_limit()}}{
Get feature limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_limit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_total"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_total}{}}}
\subsection{Method \code{get_feature_total()}}{
Get feature current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_total()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_visible}{}}}
\subsection{Method \code{get_feature_visible()}}{
Get feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_invisible}{}}}
\subsection{Method \code{get_feature_invisible()}}{
Get feature invisible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with date/time value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_loaded}{}}}
\subsection{Method \code{get_feature_loaded()}}{
Get feature loaded values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with loaded value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_hidden}{}}}
\subsection{Method \code{get_feature_hidden()}}{
Get feature hidden values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_hidden()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_status}{}}}
\subsection{Method \code{get_feature_status()}}{
Get feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} vector with status value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_goal}{}}}
\subsection{Method \code{get_feature_goal()}}{
Get feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_goal()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector with goal value(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_order}{}}}
\subsection{Method \code{get_feature_order()}}{
Set relative order for displaying features on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_feature_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_order}{}}}
\subsection{Method \code{set_feature_order()}}{
Set relative order for displaying features on a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_order(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector of new orders.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-set_loaded}{}}}
\subsection{Method \code{set_loaded()}}{
Set loaded state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_loaded(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector of loaded state.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, \code{"visible"}, or \code{"feature_order"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible value for all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_visible}{}}}
\subsection{Method \code{set_feature_visible()}}{
Set feature visible values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_status}{}}}
\subsection{Method \code{set_feature_status()}}{
Set feature status values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} vector containing a value for each feature.
A \code{list} of \code{logical} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goal}{}}}
\subsection{Method \code{set_feature_goal()}}{
Set feature goal values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_goal(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector containing a value for each feature.
A \code{list} of \code{numeric} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_current"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_current}{}}}
\subsection{Method \code{set_feature_current()}}{
Set feature current values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_feature_current(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} vector containing a value for each feature.
A \code{list} of \code{numeric} values can also be supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"feature_status"}, \code{"feature_goal"},
\code{"feature_visible"}, \code{"visible"}, \code{"feature_order"},
or \code{"feature_current"}.}

\item{\code{value}}{vector containing a value for each feature.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
