% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable-class.R
\name{Variable}
\alias{Variable}
\title{Variable class}
\description{
Variable class

Variable class
}
\details{
Definition for the Variable class.
}
\seealso{
\code{\link[=new_variable]{new_variable()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{\code{Dataset} object.}

\item{\code{index}}{\code{character} or \code{integer} value indicating the
field/layer for extracting data.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}

\item{\code{legend}}{\code{Legend} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Variable$new()}}
\item \href{#method-print}{\code{Variable$print()}}
\item \href{#method-repr}{\code{Variable$repr()}}
\item \href{#method-verify}{\code{Variable$verify()}}
\item \href{#method-get_data}{\code{Variable$get_data()}}
\item \href{#method-export}{\code{Variable$export()}}
\item \href{#method-render}{\code{Variable$render()}}
\item \href{#method-update_render}{\code{Variable$update_render()}}
\item \href{#method-clone}{\code{Variable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Variable object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$new(dataset, index, total, units, legend)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{\code{Dataset} value.}

\item{\code{index}}{\code{character} or \code{integer} value.}

\item{\code{total}}{\code{numeric} value.}

\item{\code{units}}{\code{character} value.}

\item{\code{legend}}{\code{Legend} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Variable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify"></a>}}
\if{latex}{\out{\hypertarget{method-verify}{}}}
\subsection{Method \code{verify()}}{
Verify that the data can be extracted from the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$verify()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible \code{TRUE} indicating success.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
Render variable on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$render(x, id, zindex, visible)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{id}}{\code{character} identifier for map pane.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}

\item{\code{visible}}{\code{logical} should the variable be visible?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_render"></a>}}
\if{latex}{\out{\hypertarget{method-update_render}{}}}
\subsection{Method \code{update_render()}}{
Update rendering of variable on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$update_render(x, id, zindex, visible)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{id}}{\code{character} identifier for map pane.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}

\item{\code{visible}}{\code{logical} should the variable be visible?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
