% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight-class.R
\name{Weight}
\alias{Weight}
\title{Weight class}
\description{
Weight class

Weight class
}
\details{
Definition for the Weight class.
}
\seealso{
\code{\link[=new_weight]{new_weight()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{layer}}{\link{Layer} object.}

\item{\code{status}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{factor}}{\code{numeric} value.}

\item{\code{initial_factor}}{\code{numeric} initial factor value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Weight$new()}}
\item \href{#method-get_factor}{\code{Weight$get_factor()}}
\item \href{#method-get_status}{\code{Weight$get_status()}}
\item \href{#method-set_factor}{\code{Weight$set_factor()}}
\item \href{#method-set_status}{\code{Weight$set_status()}}
\item \href{#method-get_new_solution_manager_data}{\code{Weight$get_new_solution_manager_data()}}
\item \href{#method-get_map_manager_data}{\code{Weight$get_map_manager_data()}}
\item \href{#method-clone}{\code{Weight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Weight object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$new(
  id,
  name,
  layer,
  initial_status,
  initial_factor,
  min_factor,
  max_factor,
  step_factor
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{layer}}{\link{Layer} object.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{initial_factor}}{\code{numeric} initial factor value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Weight object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor}{}}}
\subsection{Method \code{get_factor()}}{
Get factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_factor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_factor"></a>}}
\if{latex}{\out{\hypertarget{method-set_factor}{}}}
\subsection{Method \code{set_factor()}}{
Set factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_factor(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_new_solution_manager_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_new_solution_manager_data}{}}}
\subsection{Method \code{get_new_solution_manager_data()}}{
Get data for displaying the theme in a \link{newSolutionManager} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_new_solution_manager_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with data for
displaying the object in a \code{\link[=newSolutionManager]{newSolutionManager()}} widget.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_data}{}}}
\subsection{Method \code{get_map_manager_data()}}{
Get data for displaying the theme in a \link{mapManager} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_map_manager_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with data for displaying the object in a \code{\link[=mapManager]{mapManager()}}
widget.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
