% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Weight.R
\name{Weight}
\alias{Weight}
\title{Weight class}
\description{
Weight class

Weight class
}
\details{
Definition for the Weight class.
}
\seealso{
\code{\link[=new_weight]{new_weight()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} name.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{current}}{\code{numeric} value.}

\item{\code{factor}}{\code{numeric} value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Weight$new()}}
\item \href{#method-print}{\code{Weight$print()}}
\item \href{#method-repr}{\code{Weight$repr()}}
\item \href{#method-get_layer_name}{\code{Weight$get_layer_name()}}
\item \href{#method-get_layer_index}{\code{Weight$get_layer_index()}}
\item \href{#method-get_layer_pane}{\code{Weight$get_layer_pane()}}
\item \href{#method-get_id}{\code{Weight$get_id()}}
\item \href{#method-get_visible}{\code{Weight$get_visible()}}
\item \href{#method-get_invisible}{\code{Weight$get_invisible()}}
\item \href{#method-get_loaded}{\code{Weight$get_loaded()}}
\item \href{#method-get_hidden}{\code{Weight$get_hidden()}}
\item \href{#method-get_status}{\code{Weight$get_status()}}
\item \href{#method-get_current}{\code{Weight$get_current()}}
\item \href{#method-get_factor}{\code{Weight$get_factor()}}
\item \href{#method-get_data}{\code{Weight$get_data()}}
\item \href{#method-set_new_pane}{\code{Weight$set_new_pane()}}
\item \href{#method-get_setting}{\code{Weight$get_setting()}}
\item \href{#method-set_visible}{\code{Weight$set_visible()}}
\item \href{#method-set_invisible}{\code{Weight$set_invisible()}}
\item \href{#method-set_loaded}{\code{Weight$set_loaded()}}
\item \href{#method-set_status}{\code{Weight$set_status()}}
\item \href{#method-set_factor}{\code{Weight$set_factor()}}
\item \href{#method-set_current}{\code{Weight$set_current()}}
\item \href{#method-set_setting}{\code{Weight$set_setting()}}
\item \href{#method-get_solution_settings_widget_data}{\code{Weight$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Weight$get_map_manager_widget_data()}}
\item \href{#method-export}{\code{Weight$export()}}
\item \href{#method-render_on_map}{\code{Weight$render_on_map()}}
\item \href{#method-update_on_map}{\code{Weight$update_on_map()}}
\item \href{#method-clone}{\code{Weight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Weight object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$new(
  id,
  name,
  variable,
  pane,
  visible,
  invisible,
  loaded,
  hidden,
  status,
  current,
  factor,
  min_factor,
  max_factor,
  step_factor
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{variable}}{\link{Variable} object.}

\item{\code{pane}}{\code{character} value.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{invisible}}{\code{numeric} date/time value.}

\item{\code{loaded}}{\code{logical} value.}

\item{\code{hidden}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{current}}{\code{logical} value.}

\item{\code{factor}}{\code{numeric} initial factor value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Weight object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_name}{}}}
\subsection{Method \code{get_layer_name()}}{
Get layer names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_layer_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_index"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_index}{}}}
\subsection{Method \code{get_layer_index()}}{
Get layer index values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_layer_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layer_pane"></a>}}
\if{latex}{\out{\hypertarget{method-get_layer_pane}{}}}
\subsection{Method \code{get_layer_pane()}}{
Get layer pane class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_layer_pane()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get weight identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-get_invisible}{}}}
\subsection{Method \code{get_invisible()}}{
Get invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_invisible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} date/time value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-get_loaded}{}}}
\subsection{Method \code{get_loaded()}}{
Get loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_loaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-get_hidden}{}}}
\subsection{Method \code{get_hidden()}}{
Get hidden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_hidden()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_current"></a>}}
\if{latex}{\out{\hypertarget{method-get_current}{}}}
\subsection{Method \code{get_current()}}{
Get current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_current()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor}{}}}
\subsection{Method \code{get_factor()}}{
Get factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_factor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[terra:rast]{terra::rast()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_new_pane"></a>}}
\if{latex}{\out{\hypertarget{method-set_new_pane}{}}}
\subsection{Method \code{set_new_pane()}}{
Set new pane.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_new_pane(id, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} unique identifier.}

\item{\code{index}}{\code{character} variable index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} \code{"factor"}, \code{"current"}, or \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_invisible"></a>}}
\if{latex}{\out{\hypertarget{method-set_invisible}{}}}
\subsection{Method \code{set_invisible()}}{
Set invisible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_invisible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} date/time value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_loaded"></a>}}
\if{latex}{\out{\hypertarget{method-set_loaded}{}}}
\subsection{Method \code{set_loaded()}}{
Set loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_loaded(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_factor"></a>}}
\if{latex}{\out{\hypertarget{method-set_factor}{}}}
\subsection{Method \code{set_factor()}}{
Set factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_factor(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_current"></a>}}
\if{latex}{\out{\hypertarget{method-set_current}{}}}
\subsection{Method \code{set_current()}}{
Set current.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_current(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"} \code{"factor"}, \code{"current"},
or `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$render_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-update_on_map}{}}}
\subsection{Method \code{update_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$update_on_map(x, zindex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.}

\item{\code{zindex}}{\code{numeric} z-index for ordering.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
