% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight-class.R
\name{Weight}
\alias{Weight}
\title{Weight class}
\description{
Weight class

Weight class
}
\details{
Definition for the Weight class.
}
\seealso{
\code{\link[=new_weight]{new_weight()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{dataset}}{\link{Dataset} object.}

\item{\code{visible}}{\code{logical} value.}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{factor}}{\code{numeric} value.}

\item{\code{initial_factor}}{\code{numeric} initial factor value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Weight$new()}}
\item \href{#method-print}{\code{Weight$print()}}
\item \href{#method-repr}{\code{Weight$repr()}}
\item \href{#method-get_visible}{\code{Weight$get_visible()}}
\item \href{#method-get_status}{\code{Weight$get_status()}}
\item \href{#method-get_factor}{\code{Weight$get_factor()}}
\item \href{#method-get_data}{\code{Weight$get_data()}}
\item \href{#method-get_parameter}{\code{Weight$get_parameter()}}
\item \href{#method-set_visible}{\code{Weight$set_visible()}}
\item \href{#method-set_status}{\code{Weight$set_status()}}
\item \href{#method-set_factor}{\code{Weight$set_factor()}}
\item \href{#method-set_parameter}{\code{Weight$set_parameter()}}
\item \href{#method-get_solution_settings_widget_data}{\code{Weight$get_solution_settings_widget_data()}}
\item \href{#method-get_map_manager_widget_data}{\code{Weight$get_map_manager_widget_data()}}
\item \href{#method-clone}{\code{Weight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Weight object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$new(
  id,
  name,
  dataset,
  initial_visible,
  initial_status,
  initial_factor,
  min_factor,
  max_factor,
  step_factor
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{dataset}}{\link{Dataset} object.}

\item{\code{initial_visible}}{\code{logical} value.}

\item{\code{initial_status}}{\code{logical} value.}

\item{\code{initial_factor}}{\code{numeric} initial factor value.}

\item{\code{min_factor}}{\code{numeric} minimum factor value.}

\item{\code{max_factor}}{\code{numeric} maximum factor value.}

\item{\code{step_factor}}{\code{numeric} step factor value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Weight object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the parameter list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_visible"></a>}}
\if{latex}{\out{\hypertarget{method-get_visible}{}}}
\subsection{Method \code{get_visible()}}{
Get visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_visible()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor}{}}}
\subsection{Method \code{get_factor()}}{
Get factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_factor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[raster:raster]{raster::raster()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_parameter(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are \code{"status"} \code{"factor"}, or \code{"visible"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_visible"></a>}}
\if{latex}{\out{\hypertarget{method-set_visible}{}}}
\subsection{Method \code{set_visible()}}{
Set visible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_visible(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_factor"></a>}}
\if{latex}{\out{\hypertarget{method-set_factor}{}}}
\subsection{Method \code{set_factor()}}{
Set factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_factor(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$set_parameter(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} parameter name.
Available options are \code{"status"} \code{"factor"}, or `"visible"``.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_widget_data}{}}}
\subsection{Method \code{get_solution_settings_widget_data()}}{
Get data for displaying the theme in a \code{\link[=solutionSettings]{solutionSettings()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_solution_settings_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_manager_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_manager_widget_data}{}}}
\subsection{Method \code{get_map_manager_widget_data()}}{
Get data for displaying the theme in a \code{\link[=mapManager]{mapManager()}} widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$get_map_manager_widget_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with widget data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Weight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
