% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_configuration_file.R
\name{write_configuration_file}
\alias{write_configuration_file}
\title{Write a configuration file}
\usage{
write_configuration_file(x, path, name, data_path, mode = "advanced")
}
\arguments{
\item{x}{\code{list} of \link{Theme} and \link{Weight} objects.}

\item{path}{\code{character} file path to save the configuration file.}

\item{name}{\code{character} name for the scenario.}

\item{data_path}{\code{character} file path for the dataset that this
file should accompany.}

\item{mode}{\code{character} mode for running the application.
Defaults to \code{"advanced"}.}
}
\value{
Invisible \code{TRUE} indicating success.
}
\description{
Save a configuration file to disk.
The configuration file will contain all the data and parameters related
to a set of \link{Theme} and \link{Weight} objects.
}
\examples{
# TODO

}
