% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_io.R
\name{write_spatial_data}
\alias{write_spatial_data}
\title{Write spatial data}
\usage{
write_spatial_data(x, path, name = tools::file_path_sans_ext(basename(path)))
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} or \code{\link[raster:stack]{raster::stack()}} object.}

\item{path}{\code{character} file path to save zip archive.}

\item{name}{\code{character} name of the spatial data inside the zip archive.
Defaults to the file name of \code{path}.}
}
\value{
Invisible \code{TRUE} indicating success.
}
\description{
Write spatial data to disk as a zip archive.
}
\details{
This function saves the spatial dataset as a zip archive.
Vector data are saved in ESRI Shapefile format and raster data
are saved as a GeoTIFF with layer names stored in a plain text file.
}
