% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_raster_io.R
\name{writeNamedRaster}
\alias{writeNamedRaster}
\title{Write a Raster object to disk}
\usage{
writeNamedRaster(x, filename, ...)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} object.}

\item{filename}{\code{character} file path.}

\item{...}{arguments passed to \code{\link[raster:writeRaster]{raster::writeRaster()}}.}
}
\value{
None.
}
\description{
Save a \code{\link[terra:rast]{terra::rast()}} object to disk, along with a text file containing
the names of each layer in the object.
}
\details{
This file saves a raster object to disk along with
a text file containing the layer names of the raster.
}
